@@ -2,12 +2,11 @@
   "name": "darshana-green-routes-backend",
   "version": "1.0.0",
   "description": "Production-ready Green Route Planner backend with Express, MongoDB, and sustainability calculations",
-  "type": "module",
   "main": "dist/index.js",
   "scripts": {
     "dev": "tsx watch src/index.ts",
     "build": "tsc",
-    "start": "node dist/index.js",
+    "start": "node dist/src/index.js",
     "test": "jest",
     "lint": "eslint src --ext .ts"
   },
@@ -25,8 +24,8 @@
     "axios": "^1.6.2",
     "bcryptjs": "^3.0.3",
     "cors": "^2.8.5",
-    "dotenv": "^16.3.1",
-    "express": "^4.18.2",
+    "dotenv": "^16.6.1",
+    "express": "^4.21.2",
     "express-rate-limit": "^8.2.1",
     "joi": "^17.11.0",
     "jsonwebtoken": "^9.0.2",
@@ -38,15 +37,22 @@
     "winston": "^3.11.0"
   },
   "devDependencies": {
-    "@types/cors": "^2.8.16",
-    "@types/express": "^4.17.20",
+    "@rollup/plugin-commonjs": "^29.0.0",
+    "@rollup/plugin-node-resolve": "^16.0.3",
+    "@types/cors": "^2.8.19",
+    "@types/express": "^4.17.25",
     "@types/jest": "^29.5.7",
-    "@types/node": "^20.8.0",
+    "@types/node": "^20.19.25",
     "@typescript-eslint/eslint-plugin": "^6.7.5",
     "@typescript-eslint/parser": "^6.7.5",
     "eslint": "^8.51.0",
     "jest": "^29.7.0",
+    "rollup": "^4.53.3",
+    "rollup-plugin-peer-deps-external": "^2.2.4",
+    "rollup-plugin-typescript2": "^0.36.0",
+    "ts-node": "^10.9.2",
     "tsx": "^4.1.0",
-    "typescript": "^5.2.2"
-  }
+    "typescript": "^5.9.3"
+  },
+  "types": "./dist/index.d.ts"
 }
