@@ -1,4 +1,4 @@
-import mongoose from 'mongoose';
+const mongoose = require('mongoose');
 
 const chatHistorySchema = new mongoose.Schema({
   userId: {
@@ -20,5 +20,5 @@ const chatHistorySchema = new mongoose.Schema({
 });
 
 const ChatHistory = mongoose.model('ChatHistory', chatHistorySchema);
-export default ChatHistory;
+module.exports = ChatHistory;
 
