@@ -1,24 +1,24 @@
-import express from 'express';
-import cors from 'cors';
-import dotenv from 'dotenv';
-import connectDB from './config/database.js';
-import questionRoutes from './routes/questions.js';
-import chatRoutes from './routes/chat.js';
-import authRoutes from './routes/auth.js';
-import reviewRoutes from './routes/reviews.js';
-import plannerRoutes from './routes/planner.js';
-import safetyRoutes from './routes/safety.js';
-import yatraShayakRoutes from './routes/yatraShayak.js';
-import guideRegistrationRoutes from './routes/guideRegistration.js';
-import itineraryRoutes from './routes/itineraries.js';
-import aiRoutes from './routes/ai.js';
-import profileRoutes from './routes/profile.js';
-import festivalRoutes from './routes/festivals.js';
-import bookingRoutes from './routes/bookings.js';
-import contactRoutes from './routes/contact.js';
-import guideRoutes from './routes/guides.js';
-import notificationRoutes from './routes/notifications.js';
-import moodAnalyzerRoutes from './routes/moodAnalyzer.js';
+const express = require('express');
+const cors = require('cors');
+const dotenv = require('dotenv');
+const connectDB = require('./config/database');
+const questionRoutes = require('./routes/questions');
+const chatRoutes = require('./routes/chat');
+const authRoutes = require('./routes/auth');
+const reviewRoutes = require('./routes/reviews');
+const plannerRoutes = require('./routes/planner');
+const safetyRoutes = require('./routes/safety');
+const yatraShayakRoutes = require('./routes/yatraShayak');
+const guideRegistrationRoutes = require('./routes/guideRegistration');
+const itineraryRoutes = require('./routes/itineraries');
+const aiRoutes = require('./routes/ai');
+const profileRoutes = require('./routes/profile');
+const festivalRoutes = require('./routes/festivals');
+const bookingRoutes = require('./routes/bookings');
+const contactRoutes = require('./routes/contact');
+const guideRoutes = require('./routes/guides');
+const notificationRoutes = require('./routes/notifications');
+const moodAnalyzerRoutes = require('./routes/moodAnalyzer');
 
 dotenv.config();
 
