@@ -14,7 +14,7 @@ export const connectDatabase = async () => {
     return mongoose.connection;
   } catch (error) {
     logger.error('❌ MongoDB connection failed:', error);
-    process.exit(1);
+    throw error;
   }
 };
 
