@@ -3,7 +3,7 @@ import Stripe from 'stripe';
 import { environment } from './environment.js';
 
 export const stripeClient = environment.stripe.secretKey
-  ? new Stripe(environment.stripe.secretKey, { apiVersion: '2023-10-16' })
+  ? new Stripe(environment.stripe.secretKey, { apiVersion: '2023-08-16' })
   : null;
 
 export const razorpayClient =
