@@ -1,6 +1,7 @@
 import React from "react";
 import { HashRouter, Routes, Route, useLocation } from "react-router-dom";
 import { SpeedInsights } from "@vercel/speed-insights/react";
+import { Analytics } from "@vercel/analytics/react";
 
 import { AuthProvider } from "./context/AuthContext";
 import Navbar from "./components/Navbar";
@@ -16,7 +17,6 @@ import TravelHub from "./pages/TravelHub"; // ✔️ Correct Import
 import Festivals from "./pages/Festivals";
 import Sustainable from "./pages/Sustainable";
 import Assistant from "./pages/Assistant";
-import Register from "./pages/Register";
 import Login from "./pages/Login";
 import Profile from "./pages/Profile";
 import MyTrips from "./pages/MyTrips";
@@ -33,6 +33,12 @@ import Booking from "./pages/Booking";
 import YatraShayak from "./components/YatraShayak";
 import AdminDashboard from "./pages/AdminDashboard";
 import NotAuthorized from "./pages/NotAuthorized";
+import GreenRoutePlanner from "./pages/GreenRoutePlanner";
+import SafetyDashboard from "./pages/SafetyDashboard";
+import SafetyGuide from "./pages/SafetyGuide";
+import SafetyModal from "./components/SafetyModal";
+import LoginOverlay from "./components/Auth/LoginOverlay";
+
 
 // Auto scroll to top when route changes
 const ScrollToTop: React.FC = () => {
@@ -45,8 +51,9 @@ const ScrollToTop: React.FC = () => {
   return null;
 };
 
-
 const App: React.FC = () => {
+  const [isSafetyOpen, setIsSafetyOpen] = React.useState(false);
+
   return (
     <AuthProvider>
       <HashRouter>
@@ -63,10 +70,16 @@ const App: React.FC = () => {
           <Routes>
             {/* Public Routes */}
             <Route path="/" element={<Home />} />
+            <Route path="/login" element={<GuestOnly><Login /></GuestOnly>} />
+            <Route path="/register" element={<GuestOnly><Login /></GuestOnly>} />
+            <Route path="/home" element={<Home />} />
             <Route path="/mood" element={<MoodAnalyzer />} />
             <Route path="/travelhub" element={<TravelHub />} />
             <Route path="/festivals" element={<Festivals />} />
             <Route path="/sustainable" element={<Sustainable />} />
+            <Route path="/green-route-planner" element={<GreenRoutePlanner />} />
+            <Route path="/safety" element={<SafetyDashboard />} />
+            <Route path="/safety-guide" element={<SafetyGuide />} />
             <Route path="/assistant" element={<Assistant />} />
             <Route path="/festival-alerts" element={<FestivalAlerts />} />
             <Route path="/language" element={<LanguageSelector />} />
@@ -75,8 +88,6 @@ const App: React.FC = () => {
             <Route path="/not-authorized" element={<NotAuthorized />} />
             
             {/* Guest Only Routes (redirect if logged in) */}
-            <Route path="/register" element={<GuestOnly><Register /></GuestOnly>} />
-            <Route path="/login" element={<GuestOnly><Login /></GuestOnly>} />
             <Route path="/forgot-password" element={<GuestOnly><ForgotPassword /></GuestOnly>} />
             <Route path="/reset-password/:token" element={<GuestOnly><ResetPassword /></GuestOnly>} />
             
@@ -86,7 +97,7 @@ const App: React.FC = () => {
             <Route path="/rewards" element={<RequireAuth><EcoRewardsDashboard /></RequireAuth>} />
             <Route path="/become-guide" element={<RequireAuth><BecomeGuide /></RequireAuth>} />
             <Route path="/guide-dashboard" element={<RequireAuth><LocalGuideDashboard /></RequireAuth>} />
-            <Route path="/booking" element={<RequireAuth><Booking /></RequireAuth>} />
+            <Route path="/booking" element={<Booking />} />
             
             {/* Admin Only Routes */}
             <Route path="/admin" element={<RequireAdmin><AdminDashboard /></RequireAdmin>} />
@@ -94,8 +105,12 @@ const App: React.FC = () => {
         </main>
         {/* Footer */}
         <Footer />
-        <YatraShayak />
+        <YatraShayak onSafetyClick={() => setIsSafetyOpen(true)} />
         <SpeedInsights />
+        <Analytics />
+
+        <SafetyModal isOpen={isSafetyOpen} onClose={() => setIsSafetyOpen(false)} />
+        <LoginOverlay />
 
       </div>
     </HashRouter>
