@@ -1,8 +1,9 @@
 import React, { useState } from "react";
 import { Link, useLocation } from "react-router-dom";
-import { Menu, X, Leaf, Smile, Bot, Calendar, Plane, MoreVertical } from "lucide-react";
+import { Menu, X, Leaf, Smile, Calendar, Plane, MoreVertical, User } from "lucide-react";
 import RightSidebar from "./RightSidebar";
 import { useRightSidebar } from "../hooks/useRightSidebar";
+import logoImage from "../images/images-map-logo.png";
 
 const Navbar: React.FC = () => {
   const [isOpen, setIsOpen] = useState(false);
@@ -13,35 +14,43 @@ const Navbar: React.FC = () => {
     { name: "Home", path: "/" },
     { name: "Mood AI", path: "/mood", icon: Smile },
     { name: "Travel Hub", path: "/travelhub", icon: Plane },  // ✔️ FIXED ROUTE
+    // { name: "Safety", path: "/safety-guide", icon: Shield },
     { name: "Cultural Odyssey", path: "/festivals", icon: Calendar },
     { name: "Eco Travel", path: "/sustainable", icon: Leaf },
-    { name: "Assistant", path: "/assistant", icon: Bot },
+    { name: "Local-Guides", path: "/guides", icon: User },
   ];
 
   const isActive = (path: string) => location.pathname === path;
 
   return (
-    <nav className="fixed top-0 w-full bg-white/80 backdrop-blur-xl shadow-lg z-50 border-b border-white/20">
+    <nav className="fixed top-0 w-full bg-white/90 backdrop-blur-md shadow-sm z-50 border-b border-slate-100">
       <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-        <div className="flex justify-between items-center h-20">
+        <div className="flex justify-between items-center h-24">
           
           {/* Logo */}
           <Link 
             to="/" 
-            className="text-3xl font-extrabold font-serif flex items-center gap-1 select-none group"
+            className="relative flex items-center justify-center select-none group"
           >
-            {/* Dar – Dark Orange to Amber Gradient with Glow */}
-            <span className="bg-gradient-to-r from-orange-600 via-orange-500 to-amber-400 bg-clip-text text-transparent group-hover:from-orange-500 group-hover:via-orange-400 group-hover:to-amber-300 transition-all duration-300">
-              Dar
-            </span>
-
-            {/* Shana – Deep Emerald to Light Green Gradient with Glow */}
-            <span className="bg-gradient-to-r from-emerald-700 via-emerald-600 to-green-400 bg-clip-text text-transparent group-hover:from-emerald-600 group-hover:via-emerald-500 group-hover:to-green-300 transition-all duration-300">
-              Shana
-            </span>
+            {/* Logo Image (Background) */}
+            <img 
+              src={logoImage} 
+              alt="DarShana Logo" 
+              className="h-32 w-auto object-contain drop-shadow-md group-hover:scale-110 transition-transform duration-500"
+            />
+            
+            {/* Text (Overlay) */}
+            <div className="absolute inset-0 flex items-center justify-center z-10 pt-2">
+              <div className="text-3xl font-extrabold font-serif tracking-wide drop-shadow-xl">
+                <span className="bg-gradient-to-r from-orange-700 via-orange-600 to-amber-600 bg-clip-text text-transparent">
+                  Dar
+                </span>
+                <span className="bg-gradient-to-r from-emerald-800 via-emerald-700 to-green-700 bg-clip-text text-transparent">
+                  Shana
+                </span>
+              </div>
+            </div>
           </Link>
-
-
 
           {/* Desktop Menu */}
           <div className="hidden md:flex items-center space-x-1">
@@ -51,23 +60,23 @@ const Navbar: React.FC = () => {
                 <Link
                   key={link.name}
                   to={link.path}
-                  className={`px-4 py-2 rounded-lg text-sm flex items-center gap-2 font-semibold transition-all duration-300 transform hover:scale-105 ${
+                  className={`px-3 py-2 rounded-full text-sm flex items-center gap-1.5 font-medium transition-all duration-300 transform hover:scale-105 ${
                     isActive(link.path)
                       ? "text-white bg-gradient-to-r from-orange-500 to-orange-600 shadow-md"
-                      : "text-slate-700 hover:text-orange-600 hover:bg-orange-50/60"
+                      : "text-slate-600 hover:text-orange-600 hover:bg-orange-50"
                   }`}
                 >
-                  {Icon && <Icon size={18} />}
+                  {Icon && <Icon size={16} />}
                   {link.name}
                 </Link>
               );
             })}
 
             {/* CTA + Kebab Menu */}
-            <div className="ml-6 flex items-center gap-4">
+            <div className="ml-4 flex items-center gap-3">
               <Link
-                to="/register"
-                className="bg-gradient-to-r from-emerald-600 to-teal-600 hover:from-emerald-700 hover:to-teal-700 text-white px-6 py-2.5 rounded-full text-sm font-bold transition-all duration-300 shadow-lg hover:shadow-xl hover:scale-105 transform"
+                to="/booking"
+                className="bg-gradient-to-r from-emerald-600 to-teal-600 hover:from-emerald-700 hover:to-teal-700 text-white px-5 py-2 rounded-full text-sm font-bold transition-all duration-300 shadow-md hover:shadow-lg hover:scale-105 transform"
               >
                 Book Trip
               </Link>
@@ -75,10 +84,10 @@ const Navbar: React.FC = () => {
               {/* Kebab Menu (3-dot) */}
               <button
                 onClick={openSidebar}
-                className="p-2.5 text-slate-600 hover:text-orange-600 hover:bg-orange-100/40 rounded-lg transition-all duration-300 hover:scale-110"
+                className="p-2 text-slate-600 hover:text-orange-600 hover:bg-orange-100/50 rounded-full transition-all duration-300 hover:scale-110"
                 title="More options"
               >
-                <MoreVertical size={22} />
+                <MoreVertical size={20} />
               </button>
             </div>
           </div>
@@ -89,7 +98,7 @@ const Navbar: React.FC = () => {
             className="md:hidden p-2 text-gray-700 hover:text-orange-600"
             aria-label="Toggle menu"
           >
-            {isOpen ? <X size={26} /> : <Menu size={26} />}
+            {isOpen ? <X size={24} /> : <Menu size={24} />}
           </button>
         </div>
       </div>
@@ -119,7 +128,7 @@ const Navbar: React.FC = () => {
 
             {/* CTA button */}
             <Link
-              to="/register"
+              to="/booking"
               onClick={() => setIsOpen(false)}
               className="block text-center bg-gradient-to-r from-emerald-600 to-teal-600 hover:from-emerald-700 hover:to-teal-700 text-white py-3 rounded-lg mt-4 shadow-md font-bold"
             >
