@@ -19,7 +19,7 @@ const popularDestinations: DestinationCard[] = [
     name: 'Leh–Ladakh, Jammu & Kashmir',
     region: 'Trans-Himalaya',
     description: 'Confluence drives, high passes, and night skies over Pangong Tso.',
-    image: 'https://images.unsplash.com/photo-1500530855697-b586d89ba3ee?auto=format&fit=crop&w=900&q=80',
+    image: 'https://travelmykashmir.com/wp-content/uploads/2024/05/Pangong-Lake-Near-the-Mountains-Under-the-Blue-Sky-and-White-Clouds-Kashmir-leh-ladakh-tour-packages.webp',
     temperature: '12°C',
     weather: 'Clear Skies',
     mapLink: 'https://maps.app.goo.gl/sxkRXVRwVP3sa3mL7',
@@ -31,7 +31,7 @@ const popularDestinations: DestinationCard[] = [
     name: 'Coorg, Karnataka',
     region: 'Western Ghats',
     description: 'Coffee estate stays, rainforest hikes, and Kodava culinary tours.',
-    image: 'https://images.unsplash.com/photo-1469474968028-56623f02e42e?auto=format&fit=crop&w=900&q=80',
+    image: 'https://static.toiimg.com/photo/58374190.cms',
     temperature: '22°C',
     weather: 'Misty',
     mapLink: 'https://maps.app.goo.gl/7TXHyxQ7EZbnBo8b7',
@@ -43,7 +43,7 @@ const popularDestinations: DestinationCard[] = [
     name: 'Sohra & Mawlynnong, Meghalaya',
     region: 'Northeast India',
     description: 'Living root bridges, waterfalls, and clean village walks in the clouds.',
-    image: 'https://images.unsplash.com/photo-1506744038136-46273834b3fb?auto=format&fit=crop&w=900&q=80',
+    image: 'https://d34vm3j4h7f97z.cloudfront.net/optimized/4X/4/5/4/454db32ff82609d46db45be4e4f441038cad6789_2_690x428.jpeg',
     temperature: '18°C',
     weather: 'Partly Cloudy',
     mapLink: 'https://maps.app.goo.gl/6nX8pVJx8Wx9pddY8',
