@@ -1,22 +1,22 @@
 const galleryItems = [
   {
     id: 'pangong-stars',
-    media: 'https://images.unsplash.com/photo-1500530855697-b586d89ba3ee?auto=format&fit=crop&w=900&q=80',
+    media: 'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcT8TBotCIusFDxmRvkp3dGfa-pCe6Gw59i9Zg&s',
     label: 'Starlit Pangong Lake – Ladakh',
   },
   {
     id: 'varanasi-ghats',
-    media: 'https://images.unsplash.com/photo-1469474968028-56623f02e42e?auto=format&fit=crop&w=900&q=80',
+    media: 'https://static.toiimg.com/photo/msid-105562607,width-96,height-65.cms',
     label: 'Dev Deepawali Aarti – Varanasi',
   },
   {
     id: 'alleppey-dawn',
-    media: 'https://images.unsplash.com/photo-1506744038136-46273834b3fb?auto=format&fit=crop&w=900&q=80',
+    media: 'data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAAAQABAAD/2wCEAAkGBxAQDxUPEBIVEA8VEBAPEA8PEA8QEA8PFRUWFhUVFRUYHSggGBolGxUVITEhJSkrLi4uFx8zODMsNygtLisBCgoKDg0OFxAQGi0dHx0tLS0tKystLS0tLS0tLSsvLS0tLS8tLS0rLystNS0tLS0tLS0tLS0tKy0tLS0tKy0tLf/AABEIALcBEwMBIgACEQEDEQH/xAAbAAACAwEBAQAAAAAAAAAAAAACAwABBAUGB//EAD8QAAEEAAQEAwUFBgQHAQAAAAEAAgMRBBIhMQUTQVEGYXEiMoGRoRRCscHRFSNSgqLwQ3KS4QczU2Jjc/Ek/8QAGAEBAQEBAQAAAAAAAAAAAAAAAAECAwT/xAAnEQADAQACAgEDAwUAAAAAAAAAARECAxIhMRMEQVFhkbEUInGB8P/aAAwDAQACEQMRAD8A8y0pjSktKY1ekD2lNaUhqa1DUNDU+NZmFPYUppI1RhaWBZI3LQx6U0khtIggzK2ldKc55GhMaltKa1KOpYTGhCEYUogQCY0IAjaFKWBhEFQCIKU1CwjCEI2qECamNCBqMKNlGBEAhamBZbKkWGogFAiAWWzUIAipQBFSzSwEBEFKRAKUsLARAKAIgFlssKpXSINRZVmiC8qiblUSlh8JaU1iS1NYV62eVIe1NakNTmqU1BrSmsKS1NYqDQxyexyygprXIU1tKY0rKHJjXJ2L1ptjITAFkY5PjesU6xQeEYQNKYFtHJ+wmpgSwEbUAwIghCMBQsCajCEBGAlJGEEYQhECstmlkMI2pYciBWWzUGhGEoORgrNNQaEQSw5EHLDZYMCMBLBRhylLAwEQSw5EHrNEGIgl51edSlgxRLzq0JD4SAjapSIBe48iQbE5pSWpjVIbHNTWpDU1jkgHNKY1LDkxrlCjmprUpjk1rlIXsOjCcwLO2RMbMFYK2amlMa5Z2zBMDwr2RHlmlrk1rllDkxr08DyjU1MAWRsiOOff1pYZtM1gIgs4nRCcLJtQepaQZwq5wQsNNq8yy81XzFBDVnViRZeYq5iyym4SoxIueJEQlUgOgJUXNXPEyITKQpvEqvmLAJVixvHsNA8RyzMjkcAWsc6nEE0DXr+akFO7zEXMXJPE4qJ5rKByk8xlB1XRN71qsOJ8V4KNwa/ERhxAcAHZrBIANjQb36aqQVHo+YrXlZPGmBaS0ziwS00yRwsGtCBRHmokHZHgRIO4+YRh47j5hcQRJjYl74eOnaDh3HzCMPHcfMLjCFU6Jw1Azd9dR6KFp3GvH/39E5kLZNDIGEAuBcSAS3XKKG5rrovPMcDt+KOnduoUcZUz0DZB3B801rh3+q4EbPXTp5LS2E9PkpC9jtNeO6MPC5DIz2TBGexSDsdUPHdXzR3XKMPkfgkyhzNatvetvX9Ug7HpMK0POXO2PQnM+601rTvsoJgDV/7rzTJRvf0TY5AdPh8lIi9mekGIRjELgQsvayO9FaWw+SsFOyMQoyYUuS+Oht3+gv8AJMMNdPoVIXsdR+La3Vzg0eZAQ/tGL+Nv+oLmmIVZGnxQxAOOgFa9TrRrbp1UiLWdM8Uh/jb/AKgh/akIGsjevW1yziGB7mlhtjbNVXu5u/b80csjBuS33qAaNa33B6lTwKzoO4zCP8QfU6K28bg/jH1XPZG4tzAe0S0kOyEZb12A6fisMUeKz1lbksO1LLIvUEqeC1noP23B/wBQf1H8lP25h+sg/q/RY4IZS0Zmxg6HXMb76ClrlwQLW2GC47rIHDPmcL19AngtYt3iXDA1nvzDH0PogPinD9C4+jN1T+HNIoCMGzq6BrxW46hAzhgzgkQllC2jDNaXX5lxU8D+4DFeMcOyMvAe4gXkDQLPa709V4PjPjPF4gipDAwElrYHPYT2zOBt30HkvXeMeFsOEfyo2MILSS1ga4gG6sDZfLXtTwcuR69GvE8VnlJMk0j7q88jyNNtCVlkmJNkknqSbJQEeSqvgqcqM5pRNlWddHhmFY+OVzgSWNBaRI1lE2NQRbvgrnD14Qpl5pUSs3kosxCntGsCa1gS2omN0/3XY0ODAiYB5IMo7K44wNtPQNH5IUXicKxwsENdvfQ+q5zpWtOUnK4dCa1/NdlzwBqdFypSXSiW6IADdBtrv8ysspohc/M0ajMHEE9m1036rfhyA6iC6gDdg1fl0WL7STIwuGweDXmB+idQdISBdtaBfSlBTpc4DuPUFMZiG9/wWKVnYAa9NOoS3xSEg0OuhLvM1v8ABPJUzrNkB7/IpmY1tp53t8lgwLS28wA12F0NFpkmoaboU5mJBbJkaBRog2aZd6HTy+oTX4MNdHrmLpC06AtI5bzt6gI8hJN7lUYSS3ydY9aI/MoDXhXyU2z1q3Vr73b0C0STuaaoHVvWveDu/okQsAodtfx/VOyNu6Hy9f1VIZJ8WXjUaUPvN2ex3S/MfJdH7UxrOY45G3XtG9QSKAH5LNJCyiaGw6dr/VVjoBLHlGhHtMPQO/RZ1Y4az78iMTx7DA5S51nUAMfZAq9K/u0MXHsK15L5QxpaAA+ORtHMSSbHmFzfDWM+zYzmyPMQMMzC8B++eE5TlGmx+S9bivEODky3iIj+9hJEkrQcokaScrja8mebbSbOmkk2kZcPiIJg4xSskBFEscD861W3Kwn3heva9fj5LQ2LDTRHIcMXkENeJ8IHA0aObNY1IXA4DJLi5xhga9qTNKNS2KMn2rLqJIDR01cs65mvuaSR3WxN/i7aUP1ViNg+8gxmBgixMcRklLHRTSOuZua2Oha2iG6A80nrt0U4hhcKwGpnFwc4ZTMHaAaHQd7XJfVdmkn7N9UvI4FvS/kFbnEj2SDoQAbHn+ZXgvD3E8Q4vc95yZWgNLB7537Hb8V3sPjZZJGRgu1dWVucWOuoPb8F0+XVgWU1TvD6qRnU+v5BW+MAkAVqdKpLw0dA3/HIfm4kL0UwLxwL2ujykihqPivI4zgGYj2Dt/4ivdJborSh5p88m8PkktEVeZDQD6arBL4dfnqq3FAE+XdfUfsw6/kkSYVgJJpo31rROxn40fMcR4eeB7t7m7o16WoeBvAc1mhIqiauh5nzX0B+MwdgfaIb/wDbGPXquZP4k4c1x/el5Fi42PdfoclH5qrka9GXjH5PmMmAmaSDG6xvQJHzCte1xPifCl5LeaG6VbBew/7lanZnLpj8imuTGuWUPRB69Jk1hyIvWVr0eZAE9uZUMP23+JVhy6HBJAMVDe3PiuxemcdFC0ytiAPponsACPjT/wD9c/b7RNW+2c91mD0KbA5GJFi5inMWiG3mfirDlh5idhhncGZmsv70jgxo+JUcgR3vC8TZMbExwBaXEkEkA00ur6LmiNxJppdqdWgkfNd7hvAmsIlOI9oWQ6FzANRXva3udqTpODQZdJqIJOZz4z2ABG1CvLcry650n4O64m/ZwY8NIfu/1N/C04YZ/Wh63+i3HhUg/wCViYX+ROX6gu/BcziGPdDRka4s1BlidHJHY0Iu+9/Jc/6jTN/GkaPshI94K2YcDQkn0pc2LjkbrFDpQ5kTSfgX39Et/HI9Kq3bAyRmtOtG0fJv8hZyd/h8OHik5hYS7qeYQL9AFpx7sJOfbDma/dId+JXmWcWbW8YNAkc5x1N6AhuuyRNxkktoNsmiDnNad8tfivHr6bGtd37/ACdVyNKU9c/hPDjHQFnvJDC4/qvGDAOa94YxpbZDAKAa0HTSqHp5IcRxtoBBna1wNZQ1xNjfc1flSxT8e9kva/bS+Wz3iNAaPl2V4+DrZ9w+T8s6bMNMRqG/zNZQ9NEYwcg1c7IPLMG/0lF4e4ficfG+WJ+SIEAOllcHZutANN7ha+M+H3YbDve6VrqwsuKDQJzfLdCBHZf97m71021Tvjt0vkVpUwjAMZeeYEucTZMYt29e15D6I4ORE8HmEOFH3nBuvmPZPouZ4ec57nF1NDRYprbt16WQVi8XY9sbRypTz+c9rm75YGsZkO1WXOkH8q6ZSbiM611zWe9d4mthLs0rdNXMFGxfsuA7eSvh+O5uahTb0Oho9QaNHvovnkb3Hh7QT++me1jXGrGZ1DbXYL0/gvDiHDFgOYiaUF4unFpykjysFdeHCzYYem4erY7599kwOWISJjZF2haHxDDtlidG66I+6crgRqCD0IIGq+K8ZjmbM7nCSw4i5XPk0v8AjNWNl9rD1m4nC+WIsjcxj7FOkjEra6jKSiMcmOyPh+b+wqLwvrGF8IxOOfFCOV9AXFFyg7e8ws30102TpvBWANFsDQdBq+XLlsZtA73quvNKcfh0fIeZ5Kl9df4E4cTfKI8hLLX4qJ2Q+HR4sPVh6zB6sPXrOZrEiMSLFnRB6A3CRNw+IyPa/wDhc13Q7EHZc0SIhIgOlipwZHEOLgXuIc73nWdz6pfNWMSK+YgNglRCRYeYk4vFmMZqzC6OtEIU6vMVGRceLisZ3tvrt8wm4p4IFai77hSoU6RmaNTW589q7KsW5kuDDaBML9Lb/hP1b8iCPgFxzIaIvQbDzN/oFqwM7mPDNDnaQc2rgA3NeUHTUdQvF9Vjti/fLv8A3+j0/S7a3Pz4Pe+AeIywcOn5Ja1wljLaP7y/Z9xlEO0zX2XH/bOIc2WB0rxGTiSWBzmtJfMS4loq7zHp1W3heFEMbWvZiWFzpswjZBKwSMa1oa3Qa6utY5cBJJIZC57XXIz94yKy1rGOG1j4D5ryZ5McfnZ13xvbbwcvh08TswBOgZYt1lwaM2ze5PyW/lgUcm/+cg6H/KsU8PJlZH954mdI4EVI4cqqbplpulWe+i2WB2I/lv8AO1amqnUypNeGo0FIGBppjRttlNaervwWnCsuKR7QBT8NG0kVoZA5/QdAFhfJTT27e0PLyCbhMRcbmVo6TOacwe7Qb7NE7juufIm8+P0/k3l+f3/gDin/AA5nlkklw+Kw0we98uUvexwa8lw0APfyXk+PcLmwVRy0WvJcx7S0h4aBegNisw3Xsuc4Ev5pZ1Lml7CQBl3JA2+QXD4u/DFpzOEjgDlLnhxBPYC+uu4XTj5ORuPyv8HDXDlJv0z0fhPxTh8Ngm4UCSSYZnSiMNADi6+p1GXKFfHvET5sNMDC9oOClw7S5zd3OhdZFDSoel7heRw/FsLAS6MOc9wAe5ofbu+pNVfkl4jxO4k5I9O7zrf8qmfpEuT5Enf1NPmXTq3+x1fCt8l0jraXPNA2DQ/vssvHsFGcOZQzPO8ij7xaCbJ8qb3XCk4vOfvZbJJoC7Js6rJLinO1e8uNVqTt2XoXDpathyfNnr1h61mNjY6OFlPLWNYwZhlzmgL7dD8SvQcJjMELYtXEFxOnVzi4776kr5XLKCKA62vovh/EudhYnOJLsmpJsmiQL+AXbj41n0Z+Xsz0DZT2+qc2RcwS+aMSLpDVOo2VGJlzGypjZlHkvY6YmRtkXNEyMTrPUvY6OdRYeeFFOpex8r5imdIzK869Z4x+dXnWfMrzIDQHKw9Z86vOgNIkRCVZM6IPQlNXMWXib/3fxCvOhlpwLTsVGqhTk5l0cLxCENDHRu0uyH73fSlilwxb5juk5Vx8oJw9TgvEsUTs0cWQ+z0BNjNrZ1+8Pkt+I8XYeeSOTENkfymvDGgtaPaYGm6Fn3R8l4ghDa5a4s6daOmebefTPpeM8YcOle2Tlyxkcz3XyZfay17DXBo2OwS5fHMDAOUwPov9iaADNmYWm3tPp91fObVudoB6knr/AH+qz/T4Hz6PTcU8TsxE7JBC2HIX+657g4ODQdCNPd006q5fErR7rXXVa6gfULy2YKcxVcOEkkvQ+bdtO7ifEUjxWUAaa0L023BWV/GJy3LnIbVUCfrrquZnKqytrGV9jL5NP7ml8zju4/MpecJQCKlTNDbKAbq/I7IHSk+XoiESLlAbpRGJJJUyH0WhjBVqYoUBYqxmHp3UpeoljADZ1AokdwvoeFysYGs0aB7I8t188lOxXseFYnNCwnfKAfgumCrwdoSoxKueJUQlXQ1ToiVGJFzhKjEykNU6IkTGyrnNmTBKpC06HNUWHmKKQtPnmZTMl2pa2eUbmUzJatUDLVgpau0oGAogUrMizJSjbUtKzpc8xAsJQNnnDBZ36DuuW+Yk3+A0VSPLjZVZVy1qkJnP90pmKIRlTIslgKstXT4bwaSdjnRguc1wBaK93KXX9Pquhwvgcbh+9sOzyNPtVRaO1d1z1yZRvPHpnnQEQYuxx3AQQlnJLjebNmNjQCq+ZXPAVzvsqiPEcYrlKxGmkr2HhHgsE+HEjzle2RwBI0cRqBfX/YrHJyLCrN44+ziPHQwl7g1oLnEhrWtFlxOgAHdSeMscWOGVwJa4HcOBog+dr0P2DEYWZr2tZLyw7IffaB7Wpbvu4leZea09Vc67ehrPX2ei4N4clxDQ8Go/YzyOAbFEHA+/I4hrdup1XY8ReF/seDxOZuZzJMGBKDbcsjpCCD5gDT0Xm8Bx18LQGsieQQQ6aFsxbQ2AfbQPhaDjfiHE4s/vpPY9kCKNrYoRlzZTy2ANsZnanXVc+m3pO+Dp3ysxIwxHRKne55F6kANHoNkLXLVw6s+u2+q7peTj+g3DcPLqzbdl3YAGih6/FZg9EHrslAbQ9GJFhEiY2RUGwSIxIsYkRiRUprD0YkWMSIhIpSmzmqLLzFEB49S0Nq7UOIQKu0FqWrQHmV5kFqWlAdqWgtS0pQ7S5/dKu0MmopRgygWtTWU2vNRjQFcjtLUSgKkfSQSqJtMheBdi7aQNNj0K5t00jr+G+OuwhcKGVwNkgkh2UtHw1Wri/E4mYqVkDjJhxNIY5DXtg1btOhN15UvO5RQN9wR2QFc3jLdNd9JQ6/GZWOLCx4d7JsNv2TpusGZKaUVFazmKE1quhuctuG43LFCYGH2S8v8AedoSBdDYHQarnOBCWSjyn4YWmvR08NisTK8RMecz3ZAM1A33PZc0Bel/4fYjLxBkZaDzWyYY579gSCszdPfFWAdCRrW6848ZXFo1AcWg1V0aulEo2kg3fZq4bw6XEP5cTS5wa55oE5WNFucaF0B2C14fCQjAzTuIfJzY8PG0FwLHO9vmadCGPFEa/ArDhHOY7OKvK9uosU9pYdPRxRsbQy9LDq03Ar8ytdGxREMJct0MDW67lCHK8y6pJENQeiD1kzohItA1h6IOWUSIhIhTUHohIsoeizIDYJUQkWEPRCRQpt5iiyc1RCnDtS0KihyCtS0KiUBWpaG1EpQrUtCogDtS0CiAO1TtUKiEJlCmUKKKeCkyhTKFFEiFCFIgUtXaEDtAWBS1LQFFl7ppNkkmySSSTZJO5J7oLUzJCjLUzJdq7VoGZlYclWrtKKNzK8ySCitKKNzKw9JtXmSlNAkRcxZsysOVoNOdWHrOHIg5Cj86tIzKIDAooosmCKKKICK1SiAtRUogLtVaiiAlqWoogJalqKICWpaiiAtVaiiAilqKICWpaiigLUUUVBFdqKICWrtRRCl2patRAS1dqKIC7VhyiiAu1FFFQf/Z',
     label: 'Backwater sunrise – Alleppey',
   },
   {
     id: 'bandhavgarh-safari',
-    media: 'https://images.unsplash.com/photo-1484406566174-9da000fda645?auto=format&fit=crop&w=900&q=80',
+    media: 'data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAAAQABAAD/2wCEAAkGBxMTEhUTExMWFhUXGB0YGBgXGRseGxofGxoZGhoeHxgdHyggGh0nHRodITEjJSkrLi4uFx8zODMtNygtLisBCgoKDg0OGxAQGzUlICUtLS0yLzUtLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0vLS0tLf/AABEIALcBEwMBIgACEQEDEQH/xAAbAAACAwEBAQAAAAAAAAAAAAAEBQIDBgABB//EAD4QAAECBAQEBAQEBQMDBQAAAAECEQADITEEEkFRBSJhcROBkaGxwdHwBjJC4RQjUmLxcoKiFTOSFjRDU8L/xAAZAQADAQEBAAAAAAAAAAAAAAABAgMABAX/xAAsEQACAgEDAwMEAgIDAAAAAAAAAQIRIQMSMRNBUQQi8GFxgZGh0TKxFEJS/9oADAMBAAIRAxEAPwDJYvISwen2YBSlBUxLH4xVKxEVTZp0EePGDPRcrdjaehKUhg9vpAM1KnBCQAzmIS56mIVZvKLMMTf80MotFm1IvRb5RZ4bfZpFEtWY7QZIxH6VCp+7xngFWUIFujAdTFx5gBRQd6jb70iEoMvLmA6kaf4gqarKUtlJOoNBtpAyPHjJ74irlgkVbb6x7NSFlxYgCnpFC8QphQV2jpmeWxYg9SGvC0PYzwglhgkmmnT5xaqan8oLO9RruXhTMx3N/brQD4U9toqmcSAIUKwOmw9RUNErUFOKpb1NPZoEmzxmsRvb7/zC5XGF1AGUevvAZKlbxSOk+5OWquw3ViEuA9N9oqm49ypP5WsRr5wtDkVi1EorFA5hti7g3t8BMnGJQznNSOncbmGiaDXtUfCIyuFTSGIYDcwwwf4ZceIs0DGmtoWT045bMlqywkLFAs9yfaPeHcMXNU2QtrpG0wmFlhIBSCmLpU/KsZUuKggtRjW1fheIv1OGkjoXpsq2Jv8A0tKT+ZVdgaW3iE3g0sGiPUmNdiZaCCWFnc6QsThgUKJJYWLkfdo54eok8tlpeniuECS8IEhICWcWD97RdNShIKlJZhYq9/t4AxPHEo3JZh077xnMXxdcwgKL3p5Hr1i0dOc+SUtSEODVzMUiYQEBOWgNBUl2rvdze20Qn4YJvQDShFXGopf3jG8I4gJcwgk5HcgGC+McUMxRyqLRTozUqXAnVg42+RyqbLSCQb1ALelBA8+YoyVJCP5gqOzio9YzHik0Ji+TiZiakuAGF/rFekReqHcN4SpnW4UFMAHUSX6A2OtR8j5oUykswY1JAykUse/uIUr4tNIuWYuw/aBJqi7Zvc/frD7G+ROol/iN5s4ICf5jKNHQxI1Y7jzF48PGGSAkF7P8yCb2rWFcuQolgl/VvhDHAcFWVCjkGrOxqAQ4vf21jOMFyBSm+APE8QWsDOuxO51Le0eSuFKUAoAsdTQfGNBK4WlJoAouWAchgaFh3YhqRCZi1ixJbQBn2fpo3beF6n/kPTb/AMhaj8PrarDoSXjo6fjZgU3huzB+bbTpHQd0zbIeBbKl6xdLtHsiXSsctGkESKovQpLNvBOCSLJELikBoKkYhLfmAbTf6Qkk6wWTXcNMohiQ4t2iMzBg1B/aBMRj3SEh2u8BFZs5aBGMmaU42M1SQKFVdGUKjzq8XpXKCSXD3HztCcSyd4s/hmQVOzaU18odxXdgUmuEXfxhRVISx0avfd4qn41S6q8oFE54Kk4NRUEm5PxhtqWRNzeDyVKJsCeg0hvg+AKmJDkJfT94PVgUSlIlyw6iGUq9NTsK/KHhLMG/b7aOTU9S8KJ2afpl/wBjMr/DxBbN0t2icr8NPcqPZmeNConQs5vr2hjwvmCpZJCyhkFv6lpQogsxUlJJ842nqak3SBqaenpq2hGOAYaUkqmAki5NQD/SP6l7JDsKkpitOBCMgEpQVMJKJaQ6yAzlQDCWHNq3rGnnhCFLmEPLkJyy0bqNh3dh3U5dgytaJqVqT4mWYsDx5239qP6QknIkC5zn9Na47/PnzyQuXb/QunTAhRRNyJWKZM6SR0IBvEUcSsmmUD0rbT7MecQxTpCMJklyEnmnLTmVMULkFV0jf0hVPxKywlhTaEysoVuQ1v2jf8eE1j5/X7CvVTi8/P7Hq8UMzh8lK6d/vaBv+oNPCgCzV+UZ+bxGYnMml63Z4GPEVO7Me8TXpXwysvVLlGq4vxdRdOlyrSug3NIXjixKSgOSaubekJFYsqDqPeKQpvyv50hoenilVCy9RJuyE/EVOvUQNLJLnd4MEhZDgjtl+ccUK1tqmlfvs8dKo53YFJw7aOTt0ic1ki7biCF4VJBI97X94swvD0VzX6Gno0M5LuKovsCoUkvlcsAxIa771aLkSFTCEpGY9ocSJCEsdevt/iGCMYnM4yhg1A73F3DFttolLU8IpHTrlgXDvw6ov4qCEhquxBfYmtPOmkGyeBSUlgQw/qLEmgt9POPJmMUdz19vv4mAZMwsWYVzEs5fudYk3N8sqoxXCHshEpPLlALEcp13Jd/bSIYnEhAy1IrUvro3lCr+LDgJqrXf1iuZOY0LvE9rspuVDJWMypGU1sQDVupavbs8CTeIuQyT1BJPnvAwd6Fo4kgXPt93hlFCttlq8Wp6IDecdAqp4Nx7COh6+gn5AFSyD3iMybdt4ZYiRC4yDDxyQcsA5D3iIT0g6SikTRLEFsVC9RaDZOH5c7j9tYnIwoUpmhzL4fSlDXrpt2hd6THjFtNlGCKFBJBCqGjeRtesRxXCDULVl5cytkglg6rBT6XoYK4PwIoUSFFlLygUDee7WjSLny/A8IJUFpmO6glxlo6U3VyuzAtW14eMVvwI9SW3JkeGcATzLlqXMUgZgyCzAVIcDMQdntA/D0nOnO4Lk387+UPF4Iy1KmSqAqzULhQJcEE3LEVvSF3EJznOBlI6vqX0DdoLdxaNDE02P0rSTsdALtp5/ekeqm5an9Vun20LOFzRckEN8YsxawSDdrD2jj6eaPS6mLDioEu9xAnE5s6XPwxmS0y1JVlBd1KQSqtKJHObkvsIpXNLFgSWYDX7+sKPxRxefNWgTEeGpCQLEP1Y2taOnQhSbOT1M9zijdY6e2SXo+dSt1qJJ7hIKfMQt40o+ChA/NMAUeiprgV/tQJp7qhnKw4mpQl+ZWGKq7lISAltmq9XrCZYUsIKrGWliOilIP8AxWTCrkV8CqdilKWjDyXSE8ualLUD7CpPYXi3FSvDJlCclKv1rOZc5T9AOUbQBhVK8VKZSc0+ZzBOwUTlfZIDk7kdY0KuHYbDyk+JOGdSiMyVOpayWUEitf05rjpFFSJPOTM46TJS6Cta5ugaxd6s5dtHftAE2W1w1NXHsY1a8OlzKw6Mpf8AmKBqOil1Lvdn1ALuy/G8PD+GVlaw6ixZKf8AUo5m7XrD3f3+foC9v2+fsSkU+AiyRKekXTsGUGzsQ7aa2LEeYinDzMopu5ick0XjTGkrC5OgvFnE5KEgnUCrVu2o6H2hPOxCyLh3ctXWgP3pHk3NkCCXVdrXcxPa7tsruVUkNUYZPhvQU17UZvjC6RNTnUA7D7b2j0zMiEpJY/lprpeFcqZlWWcAnX5+sPCNpiTlTQyWurR7LnC/ofeBcRPv0DdS7RPAoASH7/B4zWAXkMXjDlISHJ120gCTiGoPWLzMUEkCgN/KsAyE8xFb/vGSVAcnYdIxIQkrAdSiwgkTMrKUXZu0U+GkIBYk5mT9YsRJzk5rA0G37wroorI/xalGliKmCJmIASHonLWgfegajN1itUogta57DT3gJcuYqoUA1KhyAXBYPe8BRTM20jp+KVmLJDaVjoAVMUCRnUO1PaPYvtRHeaHF4kE1gfOIEnqJsCe1/vpEcPOqxBEJGNIkgsgG0QEomDMJhgrfpDBOACQ+sZjUU8PkhAzE1aGqZgcNZqedoVTpRGrg9GI+sDZ1p1tEHG2Mn2PoeG4e5lpIclScw2DV9oSTJIQEBBeqpZYUdKi4Jtd6HygzxzPl5kKyqKElJrRaSAXrt8YT8Qx/hKZUtkABktnSSGI7839QeOyGktqOeU3uZashKSaBJr26toITYxWY8tcw0q9affSLsNxSZMnAlDqPNlQAfUm2nrEDOIGaYEgGYVFKXe1m9npeDHRq2w9W2kjzDSWlju5r3ZvveJokKWphfzLdwATAqsSCXzh9ASEny29IO4fJUqWrMVJS9ksmmpKtB09YnKEUslo6sm8FeJ8SSULSUqUghRS7lwaOBGe4lxBU6YqYr8yjbYCw+941uNkSzlShLuKNykHzD+ZMUoxpUkSZiDy0YkJc/wBxAr8+t4aL9tIWTe62av8ACchHhYVQC0rGUTMxJzeJKdJAsE0UBAGISlOFkDpMRT+3xaeuWDeE8YBlgEJR4WVBrQMc0lT/ANIU6CdAqFnEVF50liFiYJ8kEXzKfL/5cp84ltbY1ivAyiubMSFGWlX/AHFC4S2ZQSdGRTvn3i78RcDxAOGny5KFKllTyFAKTLlkJ8NKhoQkDaqiY0/4XlyMPhZuKmB5pUwQrQ/mlyk7qzAOQ/5W0iGIxwTg1P8A96YpiVAnOSxWoB7klqlgA1+WKq4U0SbUm0xZw5aUS+fLLUX5ZbFEsvukkE2JJL+kQw/C5eXImaQ5czCGLVLkVKb3IurcxHh/ClzVBSUFATLEwrCjUGycqjmlljmSoU/MDZgv4tiVy0goKkBX5FZQWrUGoyEPoKPQbJTcvuUuo/YYoTKQMqVpCQ9Kvq5L1JN7h3rCvieClZg5loTehzTZnpQDoAe8U8N4nmQQpKSpK8rsK2Y16GGomhLMgZjVgA+zkwsntY0VaM1i8QAoBEjYAl2PkkEP3PeO/glsFqbMXoHbe9nhpxiS3OrmI/S5CB5Cqu1fKKOGYJa05lKJA0skeTAAeUZtNGymJVSJi+VyCDV7D0+ukLDg1pXlDEvpGs4nw8JGdCw5DEPftGfn4VS1ACpNPvaGhIMsi8k5w1w3vDPPkSoPQFATRnYqJHsPWGmC4OArnotIDE6li/fRugJgmbgCsqSpAAAcKTdRqzmrCpGsCWorCoNITzUEgXJIJHvpHYXBkKUS+g9WPoxeGwwC2TyBQAy0NRZiX+TRTisKUCbmTNWweSqVMDAuQCUn86SQ1LM0bTW7Fm1J7c0RmI5ki4FbaW+MMuH4ChURue529hEpcogJWsodbZUhVH75WbrBWHRMSAHADEqJLlR6bV9o5NRtI7NOpMEEkLcuA5v20+MJuLLRKSqhKiMouepPqR7Q+kzGpuS7aEnTvAeOw6luyS5e/W+pHptB03Us8Gmk445MUpClc2U1jo1MrgxYPsB+ZOlNDHR19dHH0vqKESidD6tBMrhC1WY9iCfv1hxw3BEBwC+5DfQw5kyMrKKX3Nj6/V/KBKdEpMG4bgmSKWv/AIP0EXzkEUuDt9ILmhhmHMNQaKHmNIrCgq1dwb/vElO2Kp3gCVh6WcbfSF2JkZRuND8jGn8MEU/cd4RcXQUmn+YdKx7oFwuKWgch2PmIo43xBU0MoBKszhQ0rAkifzZfT/PvBipGYu4frFFJxwBxUsic+OnOUTAc1CpJqASAfJqGIYfBTpjrNg1yaktTr13pDHEoy0b5wX+Hv52HKSeZK6+tfQH2iktX2WS2VIr4Xw9lAI1uotVvl7RsMPJyZUAUSAouKElyQ+4Af/d0gXCSWUwZqD0BJ8v2hjxOVPRh0oS+dQMxwzpKiSAxoWFOod45nLe8l4NJUjzETZSyCMxyKupO9MoOpjHcVWRNUUoUGuFXNfv2hnjOJrEsBa8syhUkAhII0HLl8gW6xSnPPITLmfzMwdTAsBWzkXFHvZjBgtoZvcB4LiakrJS6ioFLJo+ZLi4YgtVKgRSHvCsChwnFeJLxTZZZW6ciRVJRuRt+mFHFMYUGSSASmYp1LJJK1XzGmumm0T/En4jXPxK1pyZbSyVqIS4/MA+XO2rUrF2m+CKdZYx4bx4zpEqVMUP4jDT5jAtzhdlgWVlU9b1feLOMzDKQaByltS1K+ZZzo6oR/hbCy0Y2UFfqcJUUgjMQWdVSmuvV94Y8cXMmKKygolpJD5gxUksSC9gQznygakZSa8BhKMF9QPH4mZ4FVOVy+ZCSr8iU5koJFEmgSWDVNasbMQZsyVOEz80tYOQBhWVKJIcks+53hBxSYyQFILL/AFklNKUdKqjZxbvHT+LrSFhE0TFTSyqMlIygAVvcikWcU1bIptPBUjFhJOuZRbroSenX0hjwjjiUEiaQFOWrWzJd9WGpjPpwxKglLlWpAqG0HX2SIMl4dCUFwbtyKLg7f3qu7sBraIyjHuXi5cobzMSqfMSlI5XYHT7/AH3jXq4LJKQkhghgVZilTsDp5X3jEeDPyoMlZuxYgLA2CwAkndqxteC4JQyiYMpblAchAUpsxVbMpm6MS8Tlp2vax1PNSMjxqRkxSJaDmdT1WTSjuXJAHfeL8aiZLP5Qc2gYkuHVy3buGjYcZw8nDKClCWjlOZrpQkZ1rDByQLD+paYVIMuZhjJUEpWFTJgKF8pyqUySAoLUoBicwr8GUW0ibaTdMTzpZyeIpAUGFyFEVaqAaE2D2ZoY4HCKy5lZQGYBg7aC3N5wJmElGWhWlW5DpIZgCo0LAsLNarxbhcYRMCCdSMtyfpSJ6l7Vt45L6TjublzwNeHYArWAlctLkhPiNkKQUZkqFwTzAMCxHaGE7h6JIKhMK0oUFA5XFMxBSzvyrCCzg5a6wj8FJBdSqtmyglRD5ggS0B2NS7VYuTaDeAyvERJCZihLlyVpWkAI51LAQkoYMyc2n6Xq9W6blD2/kl1Ns8iPiPFv+2lMtKUKdYAHM5J3soNtY9Q43/UxMHiEqJQSAdA5AJyghz6x5xdQ8UB8yxfUtV2awqC8Bfw4SBLSDnUosLsCAVE9A7ekCSi6Xz6lISlEMnY6r3HW/e8RVjRUuXLAAE+Z9PeBmqELCXUnkNu6nDuKWIJppeKChKiStaiEOLXNialoVQRXffAw/jP7j7mOirDYGUpIVnUH0yj5rEewKiax5wuSAKyGbcBX/IKJPpDnDJSQ4LP5j76e0J5PjyZgzfzJRe/50tVu9Lat5Q6MwEBaGL+ix9Wg6kWcreCrEyCAzN1FvTSFZwikHMK7fe0aeVVIUmx9R06iB8VIDEg5T7GOVSpnMprdQqmz2SFixv0PXcbwLxIZhSj0bYs/ofu8djZlCkqQX2IfuxaBFYgZMr8wlpJ7pUQfYx3aStF26Es7D3O1CNxp6QRgp9auQ17kQSsAkkEb+RgCYjIcyT9PvrDNWMnRHiCXPIr7+UFfhxpalIUrLm5gbV2Y7/KApmOSo8wD7E/tEZc9QWmYhFEmrsA2tTekK4uqNy7NpKxIQPEcrSAoHIAaKAcu9wzt3hvh+NiZh0zhyiarwwsggpCUqVMy0oSRkzWD9IU4oEVozVJcADv5ecKcJx8IlS5ExGdUla0ZAoDlWh0LD0JB5W/uUYGnFSQsvYxtxjDI8CXMlFUtQkp8VCglKsy2/MspJmHmoEABwavSF+MnD+HmKWtc1ShmlpJBKSgOedKUlD7Et0FXKlYNSSc1UkFmZsvKUWd6Pqb7u+R4olQmLdvWgGt9WEPHbvquwZbtnPcZYji5W61SBOE6WUEspiUrICzlZlMAe9d3W4bhUwhhLZ93J/8AEGA8RxKYlCMoAlhycuhJr5VFesE8Kx+ZRrV79AAw83eKtbconu7DjBLMiYFtOz2zKSyEudstKbmGnG+IlacpyrYvmar0eouGpAkgKWFS1F0Ke9i9wfX3ijF8O8NDCYojbkfzUqJy1lwsGildvINiVlSWVUaW7el/aFs2XLRQ0JsB8WPt9BEl4kpAypyvqpQUfIW87CBv4NUypLPUnVu+8ZRHch9wfCIEtRd2ovKak/plI1JP6jqT/wCNeL4ISpMqgWr87flQ4KsieyRU9v6oj+FUJTiEpQ6hlUqpoGo/fT1jSYVPiBc5mUJS8o3K1EP3CUIERnujJtfPBaFSikxTw/GCWsIkywnMSlC1K/MEpDAKJ5QXHd6AmNOcUUJbFT0AEP4aEBRIFypRSWT1y+ZhLJ4QkplLX+RBnKSnVWVYUkvtlR7iIcGwwWudPnAzPDQhWUnlcoKwGNC1Gej1ie5J7vHPnmu5Sm1Xn9cWasYZGKSZipTBSQkzVrUDlzJXQKQAASkOwqAA8LVcOw6lKEkzFFJKswXLULEKaWl1VzHa5O7j8MxC8Vmm4mcJclBCjLQouAwIFGajVqe0PZOPTNZMhM0JoEol5EJA3WrKph5+UX3yyvi/JzuMVn4zD8SpOQpfKlwCVAswL7PW1oiqZLzkoKVTCGzWNKhlEUdmO7Rp+OS5ecFYWSdCpAf/AEpKfEV6VilXBJTP4Ca6rKfShS0HTlFNWGSdNUZaVxGapCk8gVVNakBRLqD2LKJdwO9IccQ4+ZKZsrDyANJZFXUpKU511JJSyvWBMbhBLmUQqWTV0lSknspJIPYwywCAtIKpi6ULkj6dIOp6jZwhdPR3vkRYTBqSnMEuTVS5lMx1LFqP6R5PkkDMSpRVRxQqF8qGAYbqsBrqNGteGBIDrULucyR1JNEx7iuEJIMycTlIsKZh/SHtL1JpmN6UPL1e7Ovp9kYuWCQV8odh4pcgbIlJFVbA61vFv/S1ZKlyWZBqo9kpBPyvWNIOFukT1DKluRwWSmwZNyTYChVqwpDHGcNRhZGebmKl1MtJZcxRICUlY6sC3YBgxL1m3gVQUTGZFJ5SkAjQlL+bx0fUsJ+D05E58iVNVKJcvKnoMySS1nJrHsbqfQ276mTwSyZXiKJ/0jVtaxTh8eJRcH+Uo+aCb02hBgDO8IELdP36RdNmqPMFABmNKHp1jpayciyjcyJwBBBdKq+f0MSnTgM+ZsoDl7enrGYwE9SUf2tY/LaIcQxkxiBtrt16RxPT95z9NqVhacSpSiiUlKNSpWj25QKmj37wlx6Ul0mfmU36ZR//ADo+sU/zci9CpKQVXYOWJGtCYZYeYZRKEoEtJB/mFypZFABYOdHLCO2KXY6GxFgzVrkUfeDJ0lwPi5EEmXLJK1lKlG2UkevKXPV4I/k5WLu1GV9Uw7CnZl8XJD1em7xXgwnxEBYGQqSFZrM6S7N8Ibz8ICSzt1b6wOMC5Y6kCzm4tttG3II5m8eQqXmLKWSoBJS5IClMTbLQA13hdxWfJnTErUKMAQEhR6XFdY8mcEUkBncVufhEVcOVmBI86u+ocRKO1O4mk5Pk1uEny04dJQwRKSxzUNAQA9v8ARkOJzUrOduVVgb0e7Wr8obcNw85CJhTMKQQARoauxBG8LMZw5S1upYr0sHsNootqyLbeOwhmz6FLgC1jrt7xfwqQXBAB7QwHA9c2tiPkIIwnCCkhTi1Gend/usaU1twLKJoeESeTmq+/wBInjJCEh0oGbciI4OgNX6iCsuYUrHDmyNNOz57xGYpMwrmBRegHTR+nbpAs7jCy6WSEsxyl1V6/tG14lwFK03ObqKDZjm+UZfF8GmJJFMqdmeoAOnSO/T1ItFFb4PODcXElYXLGZTBJDtQaaCNTwvjYWfBS6ORUsBQZQdOZ69Uq9RGc4dwZSiEoAdNTXKdTfW/tBeC4IrxCk12L7AEepYecLqbXkvp7kfQeIyAVJYcoQvDoA0KwAD6gD/dFZ4YlCZmDSeZWHXMUo/qUWSnyCUt2aFknCzkAp8UkGcgIBFighLhVxVFtQmGknEThiiZglrWMPlWQSkMVEvY6dI5MLvwdGSPDMKgYXEZ0v4g8bJYhJ5Rzb8jwZg1YpEjOicZiE5kqQoDOCmhUhTVGtbdWhZgcZNVi5qZkpKZXg5FBKnAAbK1BRn9Y9mYlQnYCWnMlAQcxf8AM6amm5hrSv8Af8E6br9fyTwqJUtK56UrXiAKBaajN+oAXBrzfCAf4E/+6xy3ektFaf7R8PWJ8OJVxCctauRAXlSS5YcoAFtz5xPgctGeWCnN4gUsJmBxLSHYPbOSz9odvl/Psha4Xz7icYwT5oEtIIQaZUK01LjKe3vEFYMIQsGU8xPM5qFAkOyaij2Gwg/FzyuWFTlqQFqyy0ScrHulQZqamE2OkLkpBSTlUaJLUO6WJKD0Ba8Uj4JvyaXgWEE6YhU2ahVjKlsyX3I1NKDVofYhImz0IouXnCSdJiwXVX/65aQSdCWEfNZJxKsiQP6srXqzf8q9xG54RxLEJWiUopCTlajkVVmYijAkUax7xz62lFe6/wAF9LUlLH8miKUTcWpJbwsIgLUdCtYzJ75UAnuRGd4TixPnTeIzSPDw0pMwI/uKFqSK6plqH+6YY7ESJh/6ilFCZYZh/wBwGWov3clMDcJxSRw7EpIGUqZdGuEZB5DKPKK6MYVgTU3XyHYLCysXLRicRi5subNSFKQgslLigAy6JaOiHBOFZ5CFlDuD/wDItNHLcooKNHRVyV8iKODGypYQnKogpbM2wGvwHnBmHwuZlzAw/SjbqesAcPly0nNVRO9/SGIXm7bb9z8oVyAolmUqVmsgfl69TEEGq85YFm/bp+8WJmi1yNB90ipZO4roCHhG7G2h+GkhiaLB0o/vFE5GQFICgD+lQFPjC5U7mZHm/Sp+kWzcclRbb1rCO1kDQLLkLcqu+jCnlaDUYJ/8RZglPDUSgRSB1H3GjEzysMxYu0GYHBgqBagrEsdKMEYAFKHDuffyhnLAUsl+Iw/KT0hVJwKgQsgnM50h9Jl+I6T09y2kHY0BLAJDZAO1VfMCCnRmrF06SGys1GPW0JcRLZwdIbzUFINX0Hf7MZ/EzVPzHW3Qwtu6M0qDMMgEAGCVyQEOBbp9IEkyCUulRCqa2cOWf5RZNxUwyGFyrK7AEiwqB01rBBWCCsKcoVZ9oswM5lMe0H4JKs0iUp3TlzE1Ioo2f6doX8X5ZqmP6iRRtdo01aF2pobrDiM9xWWQp9DSGA4hS8B4rFBXXpEoNpixhRDhKC69slIPQk5CqxAQPcKP30iGCmpQEluinbWoj0Ysc6Q36SP9pb4AesGU3Z0xjSC8VMYlY0liYn/UVK6t+uLV4gZZk7+uWJfYk5TSvT0vCyViSoeG15SW7gP51i+QFKlTJZoSokext5RPPzwU5+dxjIIOHXMepSmWrsglJL7t5dY5E2UUy1Ehc1CSAAWYPRybECKsFhc8kyi6VZszvS7619oY4OWAgy5iEkEM4Bc+dIbz9wePsLvA8NSsQU5XdhRVVC5rvFOGQUjPScCSWSapJvpqDFzplEoUc0tV0qq2rgaQoPEZUvMZKzV3SahxZqD4w6vgm2kX5pYQUy0KSTosBQH32gSXIS5fM5vtTb7eJvmWhSjR2OXluFB3elYFxMhUtPiBaiCQ9NCLw7jKhNyGngILFJfKxA2azUpBM7F5uZJZaajqWZ6d6wiTizd66H94Gnzya633duli0RcW+SyklwbCXx0FaJwu2VQ3STUdwQ46PHkvDoRNXKJH8NiEpQk3CSEhKHOhof8AjGORiT+a+bUW/a8U4j8RqH8sNltzOeupZvKDCEk6iaUk1bNifxAjDfyJmdKpfKQmoLWIPUc3nHRjlcflzOebKlFZuShVWoP1bAR0Xz4J/k7CY0WUa9Wg5Jc3p0hIqUzLP7+8cnEB6H2gJ2ZxSHRDlnYe/rFudABrf1baEP8AEqCmekFBQ1fu0ZsXawjEz3DJ5U+0BrUwoXJufpFk45mFWFhv5R5hpBWoJTGTFCuHY9qGNFhcS4vAIwyUpDkMEsCwGpHygeTO5umkaWn3CpDfFy3EQWopKSLN8Ytw6gpAMUzZZBESod+QmRNUgFQq1W/0l4Y4vHBTbht6gkKZtwQ8LsEtj3i3H8OE1aVJflqQwqWtGWDJ2D8V4kZy3cBLhg4DMCC9femkD4jD+KhLhIYkrUm2UOB3ID06CDMRhKuXoNrVgRKqkvygVG7Gw23h028gaSPeETMqcQop/IEFL3ckpSOj/IQPipBlZUFSlZWmTG1KjZn6/wDGHMjDpyoZiJk5J7pSKP0cwr4sHK1f1vf/AEoy+QKjGvJq9ppONSUJzTUspcxKcuUcuZKQnM4dgx/eMZi8ZmU45cvKxr3/AMw9wuKUqWlJJ5QaE1AalNqEeUIsbIHiBwwUa/B/VoffmhXHFi9cwk1oGv8AEtFM2gDlwbGDJmHCJhfmICq90Ex3FZIEtKR+kP7f5gtREVgsrF8pS/Y/KJiflKVBXm/qIXKkxISyLwm1DqTH+D4nJcJ/W1w4FLCjQeccmYagOKvqG6vGGWllZhvDHCTTmd73gbEsobfeGbCVxoTApNFFL8pTWzhi94lxriaUAJlKUolqC4p633jK4CalKll2sB8Ivl4hGd2PqYLSRk2wriAqN21uSxzd4z8mVzHv84Y4ia6wwPy69ogqRlU+sazKIxnFPKkg3SBRqZkgj40g6diUBMyWQeVakgEXTyt75qdYJwHEgEkM+YuamlgSQOgiHEJXigH+kJLf6gC0Zz7dgbO/cy4nJQ5H5fv0jlz0ptWx8/lHYrCgJU98zGA14dqgxsWNFMmcRlc2c1EAY4BTGjs1IKRId+x+sUTZDAF7w0asMrqgJagC0dFiwHP0jorZIdTJJoHGXvEhhEBy3uYEOKJo1IlLnGzxCmX3I8UmpJFoIE4kMzU3iC5BUX0j0pO0agJkgDQm2hO8NeATACSblKgmEKpwtrB+HxWQAGmlaXMamK6H6k0ehAIlgbEh3D0/zEJSEkkG6ag7nUNAisWzEGniHu/h5RSJSJxzF3f9T3GlX6xSUr4IpVyNMOpIF2qPpaKZ2KLlKVO28CT8QkPFOGmj80SlwXhmQwGKUwfXb0hphsclRZyPu0ISyiK1FGi1M5Ohtf0iXYo4LsOppKgpINS5d7NFOCkIUgnQUJ3Lh/jAUlZBcUp8RD9GFSgBItVJ6k77m0FP2k3H3C7EYg8uWgloUqjAAuSPgPWFmOnAAAmjkH0AH/JHtDkSmJG4Uk9+VvhCTGBU3Mk1I5hYU8hZ3EFGYXwBeZwsVA5VMa0qCbana+sFcXw6Zc4B3QxIfqCNDuQfaFXD1tkC3SsKdNK9WJLGKkcRXMQozJiStyyaBh3b5xTbYl0MleEpMtQSoHKQpJU5AcsOh9qwhx8zOoljzWABcP0i1GMWKFk16F/OLvGlgIWlSc+YghwKAUc94zWXZkroXS5JJbJWoq4+kAHMlwX16+8aaZjgS+ZL6cw+URRMCwSS5fcRPe+6L9GPZmOCyDUFjHLmsW0jU4rhKFuU0Nm0gBPB68wJANGZrG47xVakWTejJChogFNq0aefgAUJFAQITK4Sd/aNGaZpabXBVLxRALGhgyRxWgBQSB1pFMrhtSCbfCKpapSlFOVQ6k9e8F7WBbkNFYuWWImJFCCnvS9IsOKJIImBTIyhnsK79BtCafhQwZ/vyirwFMWIppGSVUmDc7todYtCiWYuakXgRUoijH0gWViFpLsSB0EGL4kk/md2rTrSFcWOpooMxjFc0FTsfbpB5neL+QOzRRMmqDDIb/vAQWxeZR3joOE1Iox9o6HsSgAEi8dhFqBqHrHSpmc5WBPS8XFGVRDO3SGYqD0TyOVjvePRPBUzHfSBhOI5mDtFUycTWE22PuPJuDUV2oNvrvHKkKWavTr9vHnj0FPvyiaJtXDPZ9Kw6bQkkmMOH4zw1A3a6d2/aLsTigpSsoYFThtGB/zCrISKKZRPl90gqQgjV6V3hd2KA4ZLcYkApBOYmrj4R7Lw7hR1HXpA02XbpEZdD+8aSvgbTdchpUthW1R93gWRPUTZ45WJBsD6x4hVXFIVIdsafxRBAhtguInKpJNaEHYgD9vSEEtYuqp0eLpC37wlBqzRTseFGYbAAZuhFiD1BYwHjFslKgwmaUPMOoZxvs+sAE36hj1EUYmaSAFFwn8r6fOMqM0y7FLJykKKSNKH02MclZUxIcv0+UArnbNFsrEKZ2B+cM3gVJWGKAWQFI0LmvlETw6S5BCh5j5xCVizsKR6vEE1P36QmSvt7kJ3BUs6CX6t9IWrw8xBJy/frDmXOanL7fCPCsqpRoKk1yBqL4FSeJLAYj4xOTxJbamDSSLAEDd9Y8TNIDFKT1H7h9ILrwLnyCK4gSOZJG9I9PEENZyIsmSwQ7fCBFYUev3aCkgNtFpxSakFrQPklEkuHPb7aITsMCf2aKlYPrDJITcxkkywAAoE2/xEZMgOxs76XdxAScOCLizVFq6R74S0hnSdYFfUbd9BgyWNopmykrSWd+/ygIJVSgo7b1i5OPIJGUEveBtfY25dwOZhVpPKx++sRSuaBY+m0HJxaaght4LGIlkMD7GsM5PwLtXZixPEFAMU+ojoZZUmrCOgbl4G2vyZ3A4vI/Ub/SsXIxBbyrUx0dHRJI5oyZZ/EGJfxNnAPXWOjoSkPbONY9tHR0Aay0LYPFiMUPb7+EdHQKBbOVig9X+/OLRPBsTtHR0ahkyByj1ixCwBaOjoUNloWmLUsKx0dCtDplypjRWVZo6OgBsFmhjaLvGYt0+UdHQ1C2DHFMbXMXKxA3PpHR0GgbmeS1gl39o9E1KXrfpasdHQKNZ4cQSGBicqflqQ/wB+cex0ag2UqxxzZS9NIiZ7mkdHQdqBYOnEnOQ1NDFwmcpURq0dHRmgIjMmDr7RTeOjoxgcJf8AVY2+zFxrSwvTXzjo6GYEsE5CUvUXfzrFS1ipZvePI6MlbMwfxD/WfUx0dHQ+0Q//2Q==',
     label: 'Tiger trail golden hour – Bandhavgarh',
   },
 ];
@@ -41,7 +41,10 @@ const GallerySection = () => (
     </div>
 
     <div className="mt-6 text-center">
-      <button className="px-5 py-2 rounded-full border border-[#06b6d4] text-[#06b6d4] font-semibold">
+      <button 
+        onClick={() => document.getElementById('map')?.scrollIntoView({ behavior: 'smooth' })}
+        className="px-5 py-2 rounded-full border border-[#06b6d4] text-[#06b6d4] font-semibold hover:bg-[#06b6d4]/10 transition-colors"
+      >
         Watch 360° tours
       </button>
     </div>
