@@ -9,30 +9,40 @@ const categories = [
   { id: 'pilgrimage', label: 'Pilgrimage Trails', icon: Globe2, description: 'Char Dham, Jyotirlinga, and Sufi circuits with local experts.' },
 ];
 
-const TravelCategoriesSection = () => (
-  <section id="categories" className="pt-16">
-    <div className="text-center space-y-3">
-      <p className="text-sm uppercase tracking-[0.3em] text-slate-500">Choose your vibe</p>
-      <h2 className="text-3xl font-semibold text-[#0f172a]">Browse by travel categories</h2>
-    </div>
+const TravelCategoriesSection = () => {
+  const handleCategoryClick = () => {
+    const packagesSection = document.getElementById('packages');
+    if (packagesSection) {
+      packagesSection.scrollIntoView({ behavior: 'smooth' });
+    }
+  };
 
-    <div className="mt-8 grid gap-4 md:grid-cols-3">
-      {categories.map((category) => (
-        <div
-          key={category.id}
-          className="rounded-3xl border border-slate-100 bg-white shadow-sm p-5 flex items-start gap-4"
-        >
-          <div className="rounded-2xl bg-cyan-50 text-[#06b6d4] p-3">
-            <category.icon size={20} />
+  return (
+    <section id="categories" className="pt-16">
+      <div className="text-center space-y-3">
+        <p className="text-sm uppercase tracking-[0.3em] text-slate-500">Choose your vibe</p>
+        <h2 className="text-3xl font-semibold text-[#0f172a]">Browse by travel categories</h2>
+      </div>
+
+      <div className="mt-8 grid gap-4 md:grid-cols-3">
+        {categories.map((category) => (
+          <div
+            key={category.id}
+            onClick={handleCategoryClick}
+            className="rounded-3xl border border-slate-100 bg-white shadow-sm p-5 flex items-start gap-4 cursor-pointer hover:shadow-md hover:border-cyan-100 transition-all active:scale-95"
+          >
+            <div className="rounded-2xl bg-cyan-50 text-[#06b6d4] p-3">
+              <category.icon size={20} />
+            </div>
+            <div>
+              <h3 className="text-lg font-semibold text-[#0f172a]">{category.label}</h3>
+              <p className="text-sm text-slate-500">{category.description}</p>
+            </div>
           </div>
-          <div>
-            <h3 className="text-lg font-semibold text-[#0f172a]">{category.label}</h3>
-            <p className="text-sm text-slate-500">{category.description}</p>
-          </div>
-        </div>
-      ))}
-    </div>
-  </section>
-);
+        ))}
+      </div>
+    </section>
+  );
+};
 
 export default TravelCategoriesSection;
