@@ -1,6 +1,6 @@
-import { useState, useEffect } from 'react';
+﻿import { useState } from 'react';
 import { useNavigate, useSearchParams } from 'react-router-dom';
-import { Calendar, MapPin, Users, CreditCard, CheckCircle, AlertCircle, Loader } from 'lucide-react';
+import { Calendar, MapPin, Users, CreditCard, CheckCircle, AlertCircle, Loader, UserPlus, Languages, Wallet, ShieldCheck } from 'lucide-react';
 import { useAuth } from '../context/AuthContext';
 import { bookingApi } from '../services/api';
 
@@ -17,7 +17,7 @@ interface BookingFormData {
 }
 
 const Booking = () => {
-  const { isAuthenticated, user } = useAuth();
+  const { user } = useAuth();
   const navigate = useNavigate();
   const [searchParams] = useSearchParams();
 
@@ -25,7 +25,7 @@ const Booking = () => {
     destination: searchParams.get('destination') || '',
     checkIn: searchParams.get('checkIn') || '',
     checkOut: searchParams.get('checkOut') || '',
-    guests: parseInt(searchParams.get('guests') || '1'),
+    guests: parseInt(searchParams.get('guests') || '1') || 1,
     roomType: 'standard',
     specialRequests: '',
     contactName: user?.name || '',
@@ -33,23 +33,33 @@ const Booking = () => {
     contactPhone: user?.phone || ''
   });
 
+  // New features state
+  const [addGuide, setAddGuide] = useState(false);
+  const [guideDetails, setGuideDetails] = useState({
+    language: 'Hindi',
+    gender: 'Any',
+    requirements: ''
+  });
+  const [paymentMethod, setPaymentMethod] = useState('upi');
+
   const [isLoading, setIsLoading] = useState(false);
   const [error, setError] = useState('');
   const [success, setSuccess] = useState(false);
   const [bookingId, setBookingId] = useState('');
 
-  useEffect(() => {
-    if (!isAuthenticated) {
-      navigate('/login');
-      return;
-    }
-  }, [isAuthenticated, navigate]);
-
   const handleInputChange = (e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement | HTMLSelectElement>) => {
     const { name, value } = e.target;
     setFormData(prev => ({
       ...prev,
-      [name]: name === 'guests' ? parseInt(value) : value
+      [name]: name === 'guests' ? (value === '' ? 0 : parseInt(value) || 0) : value
+    }));
+  };
+
+  const handleGuideChange = (e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement | HTMLSelectElement>) => {
+    const { name, value } = e.target;
+    setGuideDetails(prev => ({
+      ...prev,
+      [name]: value
     }));
   };
 
@@ -58,9 +68,17 @@ const Booking = () => {
     setIsLoading(true);
     setError('');
 
+    // Combine guide details and payment into special requests for backend compatibility
+    let enhancedSpecialRequests = formData.specialRequests;
+    if (addGuide) {
+      enhancedSpecialRequests += `\n[Guide Requested]: Language: ${guideDetails.language}, Gender: ${guideDetails.gender}, Req: ${guideDetails.requirements}`;
+    }
+    enhancedSpecialRequests += `\n[Payment Method]: ${paymentMethod}`;
+
     try {
       const response = await bookingApi.create({
         ...formData,
+        specialRequests: enhancedSpecialRequests,
         userId: user?.id
       });
 
@@ -79,27 +97,27 @@ const Booking = () => {
 
   if (success) {
     return (
-      <div className="min-h-screen bg-gradient-to-br from-green-50 to-blue-50 flex items-center justify-center px-4">
-        <div className="max-w-md w-full bg-white rounded-2xl shadow-xl p-8 text-center">
+      <div className="min-h-screen bg-gray-50 flex items-center justify-center px-4">
+        <div className="max-w-md w-full bg-white rounded-2xl shadow-xl p-8 text-center border-t-8 border-blue-600">
           <CheckCircle className="w-16 h-16 text-green-500 mx-auto mb-4" />
           <h2 className="text-2xl font-bold text-gray-900 mb-2">Booking Confirmed!</h2>
           <p className="text-gray-600 mb-4">
-            Your booking has been successfully created.
+            Your trip with DarShana has been successfully registered.
           </p>
-          <div className="bg-gray-50 rounded-lg p-4 mb-6">
-            <p className="text-sm text-gray-500">Booking ID</p>
-            <p className="font-mono font-bold text-lg">{bookingId}</p>
+          <div className="bg-blue-50 rounded-lg p-4 mb-6 border border-blue-100">
+            <p className="text-sm text-blue-600 font-medium">Booking ID</p>
+            <p className="font-mono font-bold text-xl text-blue-800">{bookingId}</p>
           </div>
           <div className="space-y-3">
             <button
-              onClick={() => navigate('/my-bookings')}
-              className="w-full bg-green-600 text-white py-3 px-6 rounded-lg font-semibold hover:bg-green-700 transition-colors"
+              onClick={() => navigate('/my-trips')}
+              className="w-full bg-blue-600 text-white py-3 px-6 rounded-lg font-semibold hover:bg-blue-700 transition-colors shadow-md"
             >
               View My Bookings
             </button>
             <button
               onClick={() => navigate('/')}
-              className="w-full bg-gray-200 text-gray-700 py-3 px-6 rounded-lg font-semibold hover:bg-gray-300 transition-colors"
+              className="w-full bg-white text-gray-700 border border-gray-300 py-3 px-6 rounded-lg font-semibold hover:bg-gray-50 transition-colors"
             >
               Back to Home
             </button>
@@ -110,196 +128,304 @@ const Booking = () => {
   }
 
   return (
-    <div className="min-h-screen bg-gradient-to-br from-blue-50 to-indigo-50 py-12 px-4">
-      <div className="max-w-4xl mx-auto">
-        <div className="bg-white rounded-2xl shadow-xl overflow-hidden">
-          <div className="bg-gradient-to-r from-blue-600 to-indigo-600 px-8 py-6">
-            <h1 className="text-3xl font-bold text-white">Book Your Trip</h1>
-            <p className="text-blue-100 mt-2">Complete your reservation details</p>
+    <div className="min-h-screen bg-gray-100 pb-20">
+      {/* Header Banner */}
+      <div className="bg-blue-700 h-64 w-full relative overflow-hidden">
+        <div className="absolute inset-0 bg-[url('https://images.unsplash.com/photo-1524492412937-b28074a5d7da?ixlib=rb-4.0.3&auto=format&fit=crop&w=2071&q=80')] bg-cover bg-center opacity-20 mix-blend-overlay"></div>
+        <div className="absolute inset-0 bg-gradient-to-b from-blue-600/80 to-blue-800/90"></div>
+        
+        <div className="relative z-10 max-w-4xl mx-auto px-4 h-full flex flex-col justify-center pb-10">
+          <h1 className="text-5xl font-extrabold text-white tracking-tight mb-2 drop-shadow-md font-serif">
+            DarShana
+          </h1>
+          <p className="text-xl text-blue-100 font-medium tracking-wide">
+            Darshana: A glimpse of every corner of India
+          </p>
+        </div>
+      </div>
+
+      {/* Main Form Container */}
+      <div className="max-w-3xl mx-auto px-4 -mt-20 relative z-20">
+        <form onSubmit={handleSubmit} className="bg-white rounded-xl shadow-xl overflow-hidden border-t-8 border-blue-600">
+          
+          {/* Form Header */}
+          <div className="p-8 border-b border-gray-100">
+            <h2 className="text-3xl font-bold text-gray-800">Trip Booking Registration</h2>
+            <p className="text-gray-500 mt-2">Complete your details to reserve your journey.</p>
           </div>
 
-          <form onSubmit={handleSubmit} className="p-8">
-            {error && (
-              <div className="mb-6 bg-red-50 border border-red-200 rounded-lg p-4 flex items-center">
-                <AlertCircle className="w-5 h-5 text-red-500 mr-3 flex-shrink-0" />
-                <p className="text-red-700">{error}</p>
-              </div>
-            )}
-
-            <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
-              {/* Destination */}
-              <div className="md:col-span-2">
-                <label className="block text-sm font-semibold text-gray-700 mb-2">
-                  <MapPin className="w-4 h-4 inline mr-2" />
-                  Destination
-                </label>
-                <input
-                  type="text"
-                  name="destination"
-                  value={formData.destination}
-                  onChange={handleInputChange}
-                  required
-                  className="w-full px-4 py-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-                  placeholder="Enter your destination"
-                />
-              </div>
-
-              {/* Check-in Date */}
-              <div>
-                <label className="block text-sm font-semibold text-gray-700 mb-2">
-                  <Calendar className="w-4 h-4 inline mr-2" />
-                  Check-in Date
-                </label>
-                <input
-                  type="date"
-                  name="checkIn"
-                  value={formData.checkIn}
-                  onChange={handleInputChange}
-                  required
-                  min={new Date().toISOString().split('T')[0]}
-                  className="w-full px-4 py-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-                />
-              </div>
-
-              {/* Check-out Date */}
-              <div>
-                <label className="block text-sm font-semibold text-gray-700 mb-2">
-                  <Calendar className="w-4 h-4 inline mr-2" />
-                  Check-out Date
-                </label>
-                <input
-                  type="date"
-                  name="checkOut"
-                  value={formData.checkOut}
-                  onChange={handleInputChange}
-                  required
-                  min={formData.checkIn || new Date().toISOString().split('T')[0]}
-                  className="w-full px-4 py-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-                />
-              </div>
-
-              {/* Number of Guests */}
-              <div>
-                <label className="block text-sm font-semibold text-gray-700 mb-2">
-                  <Users className="w-4 h-4 inline mr-2" />
-                  Number of Guests
-                </label>
-                <select
-                  name="guests"
-                  value={formData.guests}
-                  onChange={handleInputChange}
-                  className="w-full px-4 py-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-                >
-                  {[1,2,3,4,5,6,7,8].map(num => (
-                    <option key={num} value={num}>{num} Guest{num > 1 ? 's' : ''}</option>
-                  ))}
-                </select>
-              </div>
-
-              {/* Room Type */}
-              <div>
-                <label className="block text-sm font-semibold text-gray-700 mb-2">
-                  Room Type
-                </label>
-                <select
-                  name="roomType"
-                  value={formData.roomType}
-                  onChange={handleInputChange}
-                  className="w-full px-4 py-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-                >
-                  <option value="standard">Standard Room</option>
-                  <option value="deluxe">Deluxe Room</option>
-                  <option value="suite">Suite</option>
-                  <option value="villa">Villa</option>
-                </select>
-              </div>
-
-              {/* Contact Information */}
-              <div className="md:col-span-2">
-                <h3 className="text-lg font-semibold text-gray-900 mb-4">Contact Information</h3>
-              </div>
-
-              <div>
-                <label className="block text-sm font-semibold text-gray-700 mb-2">
-                  Full Name
-                </label>
-                <input
-                  type="text"
-                  name="contactName"
-                  value={formData.contactName}
-                  onChange={handleInputChange}
-                  required
-                  className="w-full px-4 py-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-                />
-              </div>
-
-              <div>
-                <label className="block text-sm font-semibold text-gray-700 mb-2">
-                  Email Address
-                </label>
-                <input
-                  type="email"
-                  name="contactEmail"
-                  value={formData.contactEmail}
-                  onChange={handleInputChange}
-                  required
-                  className="w-full px-4 py-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-                />
-              </div>
-
-              <div className="md:col-span-2">
-                <label className="block text-sm font-semibold text-gray-700 mb-2">
-                  Phone Number
-                </label>
-                <input
-                  type="tel"
-                  name="contactPhone"
-                  value={formData.contactPhone}
-                  onChange={handleInputChange}
-                  required
-                  className="w-full px-4 py-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-                  placeholder="+1 (555) 123-4567"
-                />
-              </div>
-
-              {/* Special Requests */}
-              <div className="md:col-span-2">
-                <label className="block text-sm font-semibold text-gray-700 mb-2">
-                  Special Requests (Optional)
-                </label>
-                <textarea
-                  name="specialRequests"
-                  value={formData.specialRequests}
-                  onChange={handleInputChange}
-                  rows={4}
-                  className="w-full px-4 py-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-                  placeholder="Any special requirements or preferences..."
-                />
-              </div>
+          {error && (
+            <div className="mx-8 mt-6 p-4 bg-red-50 border border-red-200 rounded-lg flex gap-3">
+              <AlertCircle size={24} className="text-red-600 flex-shrink-0" />
+              <p className="text-red-700">{error}</p>
             </div>
+          )}
 
-            {/* Submit Button */}
-            <div className="mt-8">
-              <button
-                type="submit"
-                disabled={isLoading}
-                className="w-full bg-gradient-to-r from-blue-600 to-indigo-600 text-white py-4 px-6 rounded-lg font-semibold hover:from-blue-700 hover:to-indigo-700 transition-all duration-200 disabled:opacity-50 disabled:cursor-not-allowed flex items-center justify-center"
-              >
-                {isLoading ? (
-                  <>
-                    <Loader className="w-5 h-5 mr-2 animate-spin" />
-                    Processing Booking...
-                  </>
-                ) : (
-                  <>
-                    <CreditCard className="w-5 h-5 mr-2" />
-                    Complete Booking
-                  </>
-                )}
-              </button>
-            </div>
-          </form>
-        </div>
+          <div className="p-8 space-y-8">
+            
+            {/* Section 1: Trip Details */}
+            <section>
+              <h3 className="text-lg font-semibold text-gray-900 mb-4 flex items-center gap-2">
+                <MapPin className="text-blue-600" size={20} /> Trip Details
+              </h3>
+              <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
+                <div className="col-span-2">
+                  <label className="block text-sm font-medium text-gray-700 mb-1">Destination</label>
+                  <input
+                    type="text"
+                    name="destination"
+                    required
+                    value={formData.destination}
+                    onChange={handleInputChange}
+                    className="w-full px-4 py-3 rounded-lg border border-gray-300 focus:ring-2 focus:ring-blue-500 focus:border-transparent transition-all"
+                    placeholder="Where do you want to go?"
+                  />
+                </div>
+                <div>
+                  <label className="block text-sm font-medium text-gray-700 mb-1">Check-in Date</label>
+                  <div className="relative">
+                    <Calendar className="absolute left-3 top-3.5 text-gray-400" size={18} />
+                    <input
+                      type="date"
+                      name="checkIn"
+                      required
+                      value={formData.checkIn}
+                      onChange={handleInputChange}
+                      className="w-full pl-10 pr-4 py-3 rounded-lg border border-gray-300 focus:ring-2 focus:ring-blue-500 focus:border-transparent"
+                    />
+                  </div>
+                </div>
+                <div>
+                  <label className="block text-sm font-medium text-gray-700 mb-1">Check-out Date</label>
+                  <div className="relative">
+                    <Calendar className="absolute left-3 top-3.5 text-gray-400" size={18} />
+                    <input
+                      type="date"
+                      name="checkOut"
+                      required
+                      value={formData.checkOut}
+                      onChange={handleInputChange}
+                      className="w-full pl-10 pr-4 py-3 rounded-lg border border-gray-300 focus:ring-2 focus:ring-blue-500 focus:border-transparent"
+                    />
+                  </div>
+                </div>
+                <div>
+                  <label className="block text-sm font-medium text-gray-700 mb-1">Guests</label>
+                  <div className="relative">
+                    <Users className="absolute left-3 top-3.5 text-gray-400" size={18} />
+                    <input
+                      type="number"
+                      name="guests"
+                      min="1"
+                      required
+                      value={formData.guests}
+                      onChange={handleInputChange}
+                      className="w-full pl-10 pr-4 py-3 rounded-lg border border-gray-300 focus:ring-2 focus:ring-blue-500 focus:border-transparent"
+                    />
+                  </div>
+                </div>
+                <div>
+                  <label className="block text-sm font-medium text-gray-700 mb-1">Room Type</label>
+                  <select
+                    name="roomType"
+                    value={formData.roomType}
+                    onChange={handleInputChange}
+                    className="w-full px-4 py-3 rounded-lg border border-gray-300 focus:ring-2 focus:ring-blue-500 focus:border-transparent"
+                  >
+                    <option value="standard">Standard</option>
+                    <option value="deluxe">Deluxe</option>
+                    <option value="suite">Suite</option>
+                  </select>
+                </div>
+              </div>
+            </section>
+
+            <hr className="border-gray-100" />
+
+            {/* Section 2: Guide Option */}
+            <section className="bg-blue-50 rounded-xl p-6 border border-blue-100">
+              <div className="flex items-center justify-between mb-4">
+                <div className="flex items-center gap-3">
+                  <div className="bg-blue-100 p-2 rounded-full text-blue-600">
+                    <UserPlus size={24} />
+                  </div>
+                  <div>
+                    <h3 className="text-lg font-bold text-gray-900">Add a Local Guide?</h3>
+                    <p className="text-sm text-gray-600">Enhance your trip with a local expert.</p>
+                  </div>
+                </div>
+                <label className="relative inline-flex items-center cursor-pointer">
+                  <input 
+                    type="checkbox" 
+                    className="sr-only peer"
+                    checked={addGuide}
+                    onChange={(e) => setAddGuide(e.target.checked)}
+                  />
+                  <div className="w-14 h-7 bg-gray-200 peer-focus:outline-none peer-focus:ring-4 peer-focus:ring-blue-300 rounded-full peer peer-checked:after:translate-x-full peer-checked:after:border-white after:content-[''] after:absolute after:top-0.5 after:left-[4px] after:bg-white after:border-gray-300 after:border after:rounded-full after:h-6 after:w-6 after:transition-all peer-checked:bg-blue-600"></div>
+                </label>
+              </div>
+
+              {/* Conditional Guide Details */}
+              {addGuide && (
+                <div className="mt-4 pt-4 border-t border-blue-200 animate-in slide-in-from-top-2 duration-300">
+                  <h4 className="text-sm font-semibold text-blue-800 mb-3 uppercase tracking-wide">Guide Preferences</h4>
+                  <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
+                    <div>
+                      <label className="block text-sm font-medium text-gray-700 mb-1">Preferred Language</label>
+                      <div className="relative">
+                        <Languages className="absolute left-3 top-3.5 text-gray-400" size={18} />
+                        <select
+                          name="language"
+                          value={guideDetails.language}
+                          onChange={handleGuideChange}
+                          className="w-full pl-10 pr-4 py-3 rounded-lg border border-gray-300 focus:ring-2 focus:ring-blue-500 focus:border-transparent"
+                        >
+                          <option value="Hindi">Hindi</option>
+                          <option value="English">English</option>
+                          <option value="Local Dialect">Local Dialect</option>
+                          <option value="French">French</option>
+                          <option value="Spanish">Spanish</option>
+                        </select>
+                      </div>
+                    </div>
+                    <div>
+                      <label className="block text-sm font-medium text-gray-700 mb-1">Guide Gender Preference</label>
+                      <select
+                        name="gender"
+                        value={guideDetails.gender}
+                        onChange={handleGuideChange}
+                        className="w-full px-4 py-3 rounded-lg border border-gray-300 focus:ring-2 focus:ring-blue-500 focus:border-transparent"
+                      >
+                        <option value="Any">No Preference</option>
+                        <option value="Male">Male</option>
+                        <option value="Female">Female</option>
+                      </select>
+                    </div>
+                    <div className="col-span-2">
+                      <label className="block text-sm font-medium text-gray-700 mb-1">Specific Requirements</label>
+                      <textarea
+                        name="requirements"
+                        value={guideDetails.requirements}
+                        onChange={handleGuideChange}
+                        rows={2}
+                        className="w-full px-4 py-3 rounded-lg border border-gray-300 focus:ring-2 focus:ring-blue-500 focus:border-transparent"
+                        placeholder="E.g., History expert, Food tour specialist..."
+                      ></textarea>
+                    </div>
+                  </div>
+                </div>
+              )}
+            </section>
+
+            <hr className="border-gray-100" />
+
+            {/* Section 3: Payment Method */}
+            <section>
+              <h3 className="text-lg font-semibold text-gray-900 mb-4 flex items-center gap-2">
+                <Wallet className="text-blue-600" size={20} /> Payment Method
+              </h3>
+              <div className="grid grid-cols-1 sm:grid-cols-3 gap-4">
+                {[
+                  { id: 'upi', label: 'UPI / GPay', icon: ShieldCheck },
+                  { id: 'card', label: 'Credit/Debit Card', icon: CreditCard },
+                  { id: 'hotel', label: 'Pay at Hotel', icon: MapPin },
+                ].map((method) => (
+                  <div 
+                    key={method.id}
+                    onClick={() => setPaymentMethod(method.id)}
+                    className={`cursor-pointer rounded-xl border-2 p-4 flex flex-col items-center justify-center gap-2 transition-all ${
+                      paymentMethod === method.id 
+                        ? 'border-blue-600 bg-blue-50 text-blue-700' 
+                        : 'border-gray-200 hover:border-blue-300 text-gray-600'
+                    }`}
+                  >
+                    <method.icon size={24} />
+                    <span className="font-semibold text-sm">{method.label}</span>
+                    {paymentMethod === method.id && (
+                      <div className="absolute top-2 right-2 text-blue-600">
+                        <CheckCircle size={16} />
+                      </div>
+                    )}
+                  </div>
+                ))}
+              </div>
+            </section>
+
+            <hr className="border-gray-100" />
+
+            {/* Section 4: Contact Info */}
+            <section>
+              <h3 className="text-lg font-semibold text-gray-900 mb-4">Contact Information</h3>
+              <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
+                <div>
+                  <label className="block text-sm font-medium text-gray-700 mb-1">Full Name</label>
+                  <input
+                    type="text"
+                    name="contactName"
+                    required
+                    value={formData.contactName}
+                    onChange={handleInputChange}
+                    className="w-full px-4 py-3 rounded-lg border border-gray-300 focus:ring-2 focus:ring-blue-500 focus:border-transparent"
+                  />
+                </div>
+                <div>
+                  <label className="block text-sm font-medium text-gray-700 mb-1">Email Address</label>
+                  <input
+                    type="email"
+                    name="contactEmail"
+                    required
+                    value={formData.contactEmail}
+                    onChange={handleInputChange}
+                    className="w-full px-4 py-3 rounded-lg border border-gray-300 focus:ring-2 focus:ring-blue-500 focus:border-transparent"
+                  />
+                </div>
+                <div>
+                  <label className="block text-sm font-medium text-gray-700 mb-1">Phone Number</label>
+                  <input
+                    type="tel"
+                    name="contactPhone"
+                    required
+                    value={formData.contactPhone}
+                    onChange={handleInputChange}
+                    className="w-full px-4 py-3 rounded-lg border border-gray-300 focus:ring-2 focus:ring-blue-500 focus:border-transparent"
+                  />
+                </div>
+                <div>
+                  <label className="block text-sm font-medium text-gray-700 mb-1">Special Requests</label>
+                  <input
+                    type="text"
+                    name="specialRequests"
+                    value={formData.specialRequests}
+                    onChange={handleInputChange}
+                    className="w-full px-4 py-3 rounded-lg border border-gray-300 focus:ring-2 focus:ring-blue-500 focus:border-transparent"
+                    placeholder="Any specific needs?"
+                  />
+                </div>
+              </div>
+            </section>
+
+          </div>
+
+          {/* Footer Actions */}
+          <div className="bg-gray-50 px-8 py-6 border-t border-gray-100 flex flex-col sm:flex-row items-center justify-between gap-4">
+            <p className="text-sm text-gray-500 text-center sm:text-left">
+              By clicking Submit, you agree to our <a href="#" className="text-blue-600 hover:underline">Terms</a> and <a href="#" className="text-blue-600 hover:underline">Privacy Policy</a>.
+            </p>
+            <button
+              type="submit"
+              disabled={isLoading}
+              className="w-full sm:w-auto bg-blue-600 text-white px-8 py-3 rounded-lg font-bold text-lg shadow-lg hover:bg-blue-700 hover:shadow-xl transition-all transform hover:-translate-y-0.5 disabled:opacity-70 disabled:cursor-not-allowed flex items-center justify-center gap-2"
+            >
+              {isLoading ? (
+                <>
+                  <Loader size={20} className="animate-spin" /> Processing...
+                </>
+              ) : (
+                'Confirm Booking'
+              )}
+            </button>
+          </div>
+        </form>
       </div>
     </div>
   );
