@@ -1,9 +1,11 @@
 import { useState, useMemo, useEffect, useRef } from 'react';
-import { Calendar, MapPin, Grid, Map as MapIcon, X, TrendingUp, Download, Share2 } from 'lucide-react';
+import { Calendar, MapPin, Grid, Map as MapIcon, X, TrendingUp, Download, Share2, Phone, MessageCircle, Info, ChevronRight, Search, ChevronDown, Clock } from 'lucide-react';
 import { MapContainer, TileLayer, Marker, Popup, useMap } from 'react-leaflet';
 import 'leaflet/dist/leaflet.css';
 import L from 'leaflet';
 import domtoimage from 'dom-to-image';
+import { motion, AnimatePresence } from 'framer-motion';
+import { Link } from 'react-router-dom';
 
 // --- Leaflet marker icon fix ---
 const DefaultIcon = new L.Icon({
@@ -271,6 +273,166 @@ const culturalHighlights = [
 
 const historicalPlaces = [
   {
+        id: 1,
+        name: "Group of Monuments at Hampi",
+        era: "Vijayanagara Empire",
+        location: "Hampi, Karnataka",
+        description: "Ruins of the capital city featuring temples and stone chariots.",
+        img: "https://pin.it/4RqJyIPsK",
+        lat: 15.335,
+        lng: 76.4629
+    },
+    {
+        id: 2,
+        name: "Meenakshi Amman Temple",
+        era: "Pandyan/Nayak Dynasties",
+        location: "Madurai, Tamil Nadu",
+        description: "Historic temple known for its 14 gopurams covered with mythological figures.",
+        img: "https://pin.it/Lb8z4LMY3",
+        lat: 9.9195,
+        lng: 78.1191
+    },
+    {
+        id: 3,
+        name: "Shore Temple",
+        era: "Pallava Dynasty",
+        location: "Mahabalipuram, Tamil Nadu",
+        description: "Granite temple complex overlooking the Bay of Bengal.",
+        img: "https://pin.it/7lJ5HHDmL",
+        lat: 12.6176,
+        lng: 80.1994
+    },
+    {
+        id: 4,
+        name: "Golconda Fort",
+        era: "Kakatiya/Qutb Shahi Dynasties",
+        location: "Hyderabad, Telangana",
+        description: "Famous fort known for its acoustics and diamonds like Koh-i-Noor.",
+        img: "https://pin.it/2B7YL2O47",
+        lat: 17.3833,
+        lng: 78.4011
+    },
+    {
+        id: 5,
+        name: "Charminar",
+        era: "Qutb Shahi Dynasty",
+        location: "Hyderabad, Telangana",
+        description: "Iconic monument and mosque built in 1591, symbol of Hyderabad.",
+        img: "https://pin.it/YwJhHdoTi",
+        lat: 17.3616,
+        lng: 78.4747
+    },
+    {
+        id: 6,
+        name: "Brihadeeswarar Temple",
+        era: "Chola Empire",
+        location: "Thanjavur, Tamil Nadu",
+        description: "Massive temple dedicated to Lord Shiva, a UNESCO World Heritage Site.",
+        img: "https://pin.it/ElBy0CN4Q",
+        lat: 10.7867,
+        lng: 79.1311
+    },
+    {
+        id: 7,
+        name: "Mysore Palace",
+        era: "Wodeyar Dynasty",
+        location: "Mysore, Karnataka",
+        description: "Royal residence known for its grand durbar hall and intricate architecture.",
+        img: "https://pin.it/A8uDrWmfQ",
+        lat: 12.2958,
+        lng: 76.6394
+    },
+    {
+        id: 8,
+        name: "Belur and Halebidu Temples",
+        era: "Hoysala Empire",
+        location: "Karnataka",
+        description: "Famed for exquisite stone carvings and architecture.",
+        img: "https://pin.it/1pi94QyU3",
+        lat: 12.9735,
+        lng: 75.6019
+    },
+    {
+        id: 9,
+        name: "Srirangam Temple",
+        era: "Medieval Period",
+        location: "Tamil Nadu",
+        description: "Largest functioning Hindu temple in the world, dedicated to Lord Ranganatha.",
+        img: "https://pin.it/4tKKrUSl9",
+        lat: 10.8737,
+        lng: 78.6937
+    },
+    {
+        id: 10,
+        name: "Sravanabelagola",
+        era: "Medieval Jain",
+        location: "Karnataka",
+        description: "Famous for the 57 feet tall monolithic statue of Gommateshwara Bahubali.",
+        img: "https://pin.it/N8c7qBFjd",
+        lat: 12.8583,
+        lng: 76.4723
+    },
+    {
+        id: 11,
+        name: "Lepakshi Temple",
+        era: "Vijayanagara Empire",
+        location: "Andhra Pradesh",
+        description: "Known for its hanging pillar and exquisite murals.",
+        img: "https://pin.it/7HnOMI2Vj",
+        lat: 14.5087,
+        lng: 77.6980
+    },
+    {
+        id: 12,
+        name: "Warangal Fort",
+        era: "Kakatiya Dynasty",
+        location: "Telangana",
+        description: "Historic fort ruins known for impressive stone gateways (Kakatiya Kala Thoranam).",
+        img: "https://pin.it/7QUakYyWm",
+        lat: 18.0063,
+        lng: 79.5832
+    },
+    {
+        id: 13,
+        name: "Nandi Hills",
+        era: "Medieval",
+        location: "Karnataka",
+        description: "Historic fort and hill station, popular for sunrise views and trekking.",
+        img: "https://pin.it/372zCWvpl",
+        lat: 13.3718,
+        lng: 77.6831
+    },
+    {
+        id: 14,
+        name: "Badami Caves",
+        era: "Chalukya Dynasty",
+        location: "Karnataka",
+        description: "Rock-cut cave temples famous for frescoes and sculptures.",
+        img: "https://pin.it/3YOAVM4oo",
+        lat: 15.9145,
+        lng: 75.6761
+    },
+    {
+        id: 15,
+        name: "Thanjavur Maratha Palace",
+        era: "Maratha Dynasty",
+        location: "Thanjavur, Tamil Nadu",
+        description: "Royal palace complex featuring museums and art galleries.",
+        img: "https://pin.it/6H1sW8iqk",
+        lat: 10.7867,
+        lng: 79.1311
+    },
+    {
+        id: 16,
+        name: "Udayagiri Caves",
+        era: "Pallava Dynasty",
+        location: "Andhra Pradesh",
+        description: "Ancient rock-cut cave temples with sculptures and inscriptions.",
+        img: "https://pin.it/6RpRlx7C5",
+        lat: 14.6437,
+        lng: 79.5776
+    },
+  {
     id: 1,
     name: "Taj Mahal",
     era: "Mughal Empire",
@@ -530,27 +692,42 @@ const NextMonthHighlight = ({ festivalsData }: { festivalsData: FestivalDataType
   );
   if (nextMonthFestivals.length === 0) {
     return (
-      <div className="bg-orange-50 border-l-4 border-orange-500 p-4 mb-8 rounded-lg shadow-sm flex items-center gap-3">
-        <TrendingUp size={20} className="text-orange-600 flex-shrink-0" />
-        <p className="text-sm text-orange-800 font-medium">
-          <b>Next Month ({nextMonthName})</b>: No major events scheduled yet. Check back for updates!
-        </p>
-      </div>
+      <motion.div 
+        initial={{ opacity: 0, y: 20 }}
+        animate={{ opacity: 1, y: 0 }}
+        className="bg-orange-50/80 backdrop-blur-sm border-l-4 border-orange-500 p-6 mb-12 rounded-xl shadow-sm flex items-center gap-4 max-w-4xl mx-auto"
+      >
+        <div className="p-3 bg-orange-100 rounded-full">
+          <TrendingUp size={24} className="text-orange-600" />
+        </div>
+        <div>
+          <h3 className="text-lg font-bold text-orange-900">Next Month ({nextMonthName})</h3>
+          <p className="text-orange-800/80">No major events scheduled yet. Check back for updates!</p>
+        </div>
+      </motion.div>
     );
   }
   const festivalNames = nextMonthFestivals.map(f => f.name).join(', ');
   return (
-    <div className="bg-green-50 border-l-4 border-green-500 p-4 mb-8 rounded-lg shadow-lg flex items-start gap-3 animate-pulse-slow">
-      <TrendingUp size={20} className="text-green-600 flex-shrink-0 mt-0.5" />
-      <div className='flex-grow'>
-        <p className="text-sm text-green-800 font-bold">
-          ✨ UPCOMING EVENTS - {nextMonthName.toUpperCase()} ✨
-        </p>
-        <p className="text-xs text-green-700 mt-1">
-          Don't miss: <b>{festivalNames}</b>. Plan your travels now!
-        </p>
+    <motion.div 
+      initial={{ opacity: 0, y: 20 }}
+      animate={{ opacity: 1, y: 0 }}
+      className="bg-gradient-to-r from-emerald-50 to-teal-50 border border-emerald-100 p-6 mb-12 rounded-2xl shadow-lg flex items-start gap-4 max-w-4xl mx-auto relative overflow-hidden"
+    >
+      <div className="absolute top-0 right-0 w-32 h-32 bg-emerald-200/20 rounded-full -mr-10 -mt-10 blur-2xl"></div>
+      <div className="p-3 bg-emerald-100 rounded-full z-10">
+        <TrendingUp size={24} className="text-emerald-600" />
       </div>
-    </div>
+      <div className='flex-grow z-10'>
+        <p className="text-sm font-bold text-emerald-600 tracking-wider uppercase mb-1">
+          ✨ Upcoming Events - {nextMonthName}
+        </p>
+        <h3 className="text-xl font-bold text-gray-900 mb-2">
+          Don't miss: <span className="text-emerald-700">{festivalNames}</span>
+        </h3>
+        <p className="text-gray-600 text-sm">Plan your travels now to witness these spectacular celebrations.</p>
+      </div>
+    </motion.div>
   );
 };
 
@@ -577,15 +754,92 @@ function LiveLocationMarker() {
   );
 }
 
+// --- HELPLINE MODAL COMPONENT ---
+const HelplineModal = ({ onClose }: { onClose: () => void }) => {
+  return (
+    <motion.div 
+      initial={{ opacity: 0 }}
+      animate={{ opacity: 1 }}
+      exit={{ opacity: 0 }}
+      className="fixed inset-0 bg-black/60 backdrop-blur-sm z-[60] flex items-center justify-center p-4"
+      onClick={onClose}
+    >
+      <motion.div 
+        initial={{ scale: 0.9, opacity: 0 }}
+        animate={{ scale: 1, opacity: 1 }}
+        exit={{ scale: 0.9, opacity: 0 }}
+        className="bg-white rounded-2xl shadow-2xl max-w-md w-full overflow-hidden"
+        onClick={e => e.stopPropagation()}
+      >
+        <div className="bg-gradient-to-r from-orange-500 to-pink-600 p-6 text-white relative">
+          <h2 className="text-2xl font-bold flex items-center gap-2">
+            <Info className="w-6 h-6" /> Cultural Helpline
+          </h2>
+          <p className="text-white/90 text-sm mt-1">Get assistance with your cultural journey</p>
+          <button 
+            onClick={onClose}
+            className="absolute top-4 right-4 p-1 bg-white/20 rounded-full hover:bg-white/30 transition"
+          >
+            <X size={20} />
+          </button>
+        </div>
+        
+        <div className="p-6 space-y-4">
+          <Link to="/assistant" className="flex items-center gap-4 p-4 bg-indigo-50 rounded-xl hover:bg-indigo-100 transition group">
+            <div className="w-12 h-12 bg-indigo-100 rounded-full flex items-center justify-center text-indigo-600 group-hover:bg-white transition">
+              <MessageCircle size={24} />
+            </div>
+            <div className="flex-1">
+              <h3 className="font-bold text-indigo-900">Ask AI Cultural Guide</h3>
+              <p className="text-xs text-indigo-700">Get instant history & tips</p>
+            </div>
+            <ChevronRight className="text-indigo-400" />
+          </Link>
+
+          <a href="tel:1363" className="flex items-center gap-4 p-4 bg-orange-50 rounded-xl hover:bg-orange-100 transition group">
+            <div className="w-12 h-12 bg-orange-100 rounded-full flex items-center justify-center text-orange-600 group-hover:bg-white transition">
+              <Phone size={24} />
+            </div>
+            <div className="flex-1">
+              <h3 className="font-bold text-orange-900">Tourist Helpline</h3>
+              <p className="text-xs text-orange-700">Call 1363 (24/7 Support)</p>
+            </div>
+            <ChevronRight className="text-orange-400" />
+          </a>
+
+          <a href="tel:112" className="flex items-center gap-4 p-4 bg-red-50 rounded-xl hover:bg-red-100 transition group">
+            <div className="w-12 h-12 bg-red-100 rounded-full flex items-center justify-center text-red-600 group-hover:bg-white transition">
+              <Phone size={24} />
+            </div>
+            <div className="flex-1">
+              <h3 className="font-bold text-red-900">Emergency</h3>
+              <p className="text-xs text-red-700">Call 112 for immediate help</p>
+            </div>
+            <ChevronRight className="text-red-400" />
+          </a>
+        </div>
+        
+        <div className="p-4 bg-gray-50 text-center text-xs text-gray-500 border-t">
+          DarShana Travel • Connecting you to India's Heritage
+        </div>
+      </motion.div>
+    </motion.div>
+  );
+};
+
 // --- MAIN COMPONENT ---
 const Festivals = () => {
   const [searchText, setSearchText] = useState('');
   const [filterType, setFilterType] = useState('all');
+  const [selectedMonth, setSelectedMonth] = useState('All Months');
+  const [upcomingFilter, setUpcomingFilter] = useState<'none' | 'next_month' | 'current_month'>('none');
   const [showMore, setShowMore] = useState(false);
   const [showMap, setShowMap] = useState(false);
   const [selectedCard, setSelectedCard] = useState<CardType | null>(null);
   const [hoveredCardId, setHoveredCardId] = useState<number | null>(null);
   const [showLiveLocation, setShowLiveLocation] = useState(false);
+  const [showHelpline, setShowHelpline] = useState(false);
+  const [isFilterOpen, setIsFilterOpen] = useState(false); // For the unified button dropdown
   const mapRef = useRef<L.Map>(null);
 
   // Create cards for all types
@@ -598,7 +852,25 @@ const Festivals = () => {
   // --- SEARCH & FILTER HOOK ---
   const filteredCards = useMemo(() => {
     let cards = allCards;
+    
+    // 1. Type Filter
     if (filterType !== 'all') cards = cards.filter(card => card.cardType === filterType);
+    
+    // 2. Month Filter
+    if (selectedMonth !== 'All Months') {
+      cards = cards.filter(card => 'month' in card && card.month.includes(selectedMonth));
+    }
+
+    // 3. Upcoming Filter
+    if (upcomingFilter === 'next_month') {
+      const next = getNextMonthName();
+      cards = cards.filter(card => 'month' in card && card.month.includes(next));
+    } else if (upcomingFilter === 'current_month') {
+      const current = months[new Date().getMonth()];
+      cards = cards.filter(card => 'month' in card && card.month.includes(current));
+    }
+
+    // 4. Search Text
     if (searchText.trim()) {
       return cards.filter(card =>
         card.name.toLowerCase().includes(searchText.toLowerCase()) ||
@@ -606,7 +878,7 @@ const Festivals = () => {
       );
     }
     return cards;
-  }, [filterType, searchText, allCards]);
+  }, [filterType, searchText, selectedMonth, upcomingFilter, allCards]);
 
   const showCards = showMore ? filteredCards : filteredCards.slice(0, 6);
   const highlightedCard = selectedCard || filteredCards[0];
@@ -630,268 +902,431 @@ const Festivals = () => {
   }
 
   return (
-    <div className="max-w-7xl mx-auto px-4 py-12 relative">
-      {/* SEARCH BAR */}
-      <div className="mb-4 flex gap-3 justify-center items-center">
-        <input
-          type="text"
-          placeholder="Search by name/location…"
-          value={searchText}
-          onChange={e => setSearchText(e.target.value)}
-          className="px-4 py-2 rounded border"
-        />
-        <button
-          className={`px-4 py-2 rounded-lg font-bold border text-sm ${showLiveLocation ? 'bg-teal-700 text-white border-teal-700' : 'bg-white border-stone-200 text-teal-800'}`}
-          onClick={() => setShowLiveLocation(x => !x)}
-        >
-          <Share2 size={16} /> Live Location
-        </button>
-        <button
-          onClick={async () => {
-            navigator.geolocation.getCurrentPosition(async ({ coords }) => {
-              const shareUrl = `${window.location.origin}?lat=${coords.latitude}&lng=${coords.longitude}`;
-              
-              if (navigator.share) {
-                try {
-                  await navigator.share({
-                    title: 'My Travel Location - DarShana',
-                    text: `Check out my current location on DarShana Travel Explorer! 📍`,
-                    url: shareUrl
-                  });
-                } catch (err) {
-                  if (err instanceof Error && err.name !== 'AbortError') {
-                    // Fallback: Copy to clipboard
-                    navigator.clipboard.writeText(shareUrl);
-                    alert('📍 Location link copied! Share it with friends.');
-                  }
-                }
-              } else {
-                // Fallback for browsers without Web Share API
-                navigator.clipboard.writeText(shareUrl);
-                alert('📍 Location link copied! Share it with friends.');
-              }
-            });
-          }}
-          className="px-4 py-2 rounded-lg font-bold bg-orange-600 text-white border border-orange-600 text-sm hover:bg-orange-700 transition"
-        >
-          <Share2 size={16} /> Share Location
-        </button>
-      </div>
-      <div className="text-center mb-8">
-        <h1 className="text-4xl font-serif font-bold text-stone-800 mb-4">🇮🇳 Indian Festival Explorer 🇮🇳</h1>
-        <p className="text-stone-600">Immerse yourself in the vibrant colors of Indian culture, religion, and history.</p>
-      </div>
-      <NextMonthHighlight festivalsData={festivalsData} />
-      {/* CATEGORY FILTER */}
-      <div className="flex gap-3 mb-8 justify-center flex-wrap">
-        {filterCategories.map(cat => (
-          <button
-            key={cat.value}
-            className={`px-4 py-2 rounded-lg font-bold border text-sm ${filterType === cat.value ? 'bg-teal-700 text-white border-teal-700' : 'bg-white text-stone-700 border-stone-200'}`}
-            onClick={() => { setFilterType(cat.value); setShowMore(false); }}
+    <div className="min-h-screen bg-stone-50 pb-20">
+      {/* HERO SECTION */}
+      <div className="relative bg-stone-900 h-[400px] flex items-center justify-center overflow-hidden">
+        <div className="absolute inset-0 opacity-60">
+          <img 
+            src="https://images.pexels.com/photos/1007426/pexels-photo-1007426.jpeg" 
+            alt="Indian Culture" 
+            className="w-full h-full object-cover"
+          />
+        </div>
+        <div className="absolute inset-0 bg-gradient-to-t from-stone-900 via-stone-900/50 to-transparent"></div>
+        
+        <div className="relative z-10 text-center px-4 max-w-4xl mx-auto">
+          <motion.h1 
+            initial={{ y: -20, opacity: 0 }}
+            animate={{ y: 0, opacity: 1 }}
+            className="text-5xl md:text-7xl font-serif font-bold text-white mb-6 tracking-tight"
           >
-            {cat.label}
-          </button>
-        ))}
+            Cultural Odyssey
+          </motion.h1>
+          <motion.p 
+            initial={{ y: 20, opacity: 0 }}
+            animate={{ y: 0, opacity: 1 }}
+            transition={{ delay: 0.2 }}
+            className="text-xl text-stone-200 font-light mb-8"
+          >
+            Immerse yourself in the vibrant colors of Indian festivals, heritage, and history.
+          </motion.p>
+          
+          {/* SEARCH BAR IN HERO */}
+          <motion.div 
+            initial={{ scale: 0.9, opacity: 0 }}
+            animate={{ scale: 1, opacity: 1 }}
+            transition={{ delay: 0.4 }}
+            className="bg-white/10 backdrop-blur-md p-2 rounded-full border border-white/20 flex items-center max-w-xl mx-auto shadow-2xl"
+          >
+            <Search className="text-white/70 ml-4 w-5 h-5" />
+            <input
+              type="text"
+              placeholder="Search festivals, places, culture..."
+              value={searchText}
+              onChange={e => setSearchText(e.target.value)}
+              className="bg-transparent border-none outline-none text-white placeholder-white/60 px-4 py-2 flex-grow w-full"
+            />
+          </motion.div>
+        </div>
       </div>
 
-      {/* LIST OR MAP VIEW */}
-      {!showMap ? (
-        <>
-          <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-8">
-            {showCards.map(card => (
-              <div
-                key={card.id || card.name}
-                className="bg-white rounded-xl shadow-lg border border-stone-100 overflow-hidden hover:shadow-2xl transition-all group cursor-pointer"
-                onMouseEnter={() => card.id && setHoveredCardId(card.id)}
-                onMouseLeave={() => setHoveredCardId(null)}
-                onClick={() => setSelectedCard(card)}
+      <div className="max-w-7xl mx-auto px-4 -mt-10 relative z-20">
+        <NextMonthHighlight festivalsData={festivalsData} />
+        
+        {/* CONTROLS BAR */}
+        <div className="flex flex-col xl:flex-row justify-between items-start xl:items-center gap-4 mb-10">
+          
+          <div className="flex flex-wrap items-center gap-3 w-full xl:w-auto">
+            {/* UNIFIED FILTER BUTTON */}
+            <div className="relative z-30">
+              <button 
+                onClick={() => setIsFilterOpen(!isFilterOpen)}
+                className="px-5 py-2.5 bg-white border border-stone-200 rounded-xl shadow-sm font-medium text-stone-700 flex items-center gap-2 min-w-[180px] justify-between hover:bg-stone-50 transition-all"
               >
-                <div className="relative h-48 overflow-hidden">
-                  <img
-                    src={card.img || 'https://placehold.co/500x300?text=Culture+or+Historical'}
-                    alt={card.name}
-                    className="w-full h-full object-cover group-hover:scale-110 transition-transform duration-500"
-                    onError={(e) => {
-                      const img = e.target as HTMLImageElement;
-                      img.src = 'https://via.placeholder.com/500x300?text=Festival+Image';
+                <span className="capitalize">{filterCategories.find(c => c.value === filterType)?.label}</span>
+                <ChevronDown size={16} className={`transition-transform ${isFilterOpen ? 'rotate-180' : ''}`} />
+              </button>
+              
+              <AnimatePresence>
+                {isFilterOpen && (
+                  <motion.div 
+                    initial={{ opacity: 0, y: 10 }}
+                    animate={{ opacity: 1, y: 0 }}
+                    exit={{ opacity: 0, y: 10 }}
+                    className="absolute top-full left-0 mt-2 w-56 bg-white rounded-xl shadow-xl border border-stone-100 overflow-hidden"
+                  >
+                    {filterCategories.map(cat => (
+                      <button
+                        key={cat.value}
+                        className={`w-full text-left px-4 py-3 text-sm font-medium hover:bg-stone-50 transition-colors flex items-center justify-between ${
+                          filterType === cat.value ? 'bg-stone-50 text-stone-900' : 'text-stone-600'
+                        }`}
+                        onClick={() => { 
+                          setFilterType(cat.value); 
+                          setIsFilterOpen(false); 
+                          setShowMore(false); 
+                        }}
+                      >
+                        {cat.label}
+                        {filterType === cat.value && <div className="w-2 h-2 rounded-full bg-teal-500"></div>}
+                      </button>
+                    ))}
+                  </motion.div>
+                )}
+              </AnimatePresence>
+            </div>
+
+            {/* MONTH SELECTION */}
+            <div className="relative group z-20">
+              <select
+                value={selectedMonth}
+                onChange={(e) => { setSelectedMonth(e.target.value); setUpcomingFilter('none'); }}
+                className="appearance-none px-5 py-2.5 bg-white border border-stone-200 rounded-xl shadow-sm font-medium text-stone-700 pr-10 hover:bg-stone-50 transition-all cursor-pointer outline-none focus:ring-2 focus:ring-stone-200"
+              >
+                <option value="All Months">All Months</option>
+                {months.map(m => <option key={m} value={m}>{m}</option>)}
+              </select>
+              <ChevronDown size={16} className="absolute right-3 top-1/2 -translate-y-1/2 text-stone-400 pointer-events-none" />
+            </div>
+
+            {/* UPCOMING BUTTONS */}
+            <div className="flex bg-white rounded-xl border border-stone-200 p-1">
+              <button
+                onClick={() => { setUpcomingFilter(upcomingFilter === 'next_month' ? 'none' : 'next_month'); setSelectedMonth('All Months'); }}
+                className={`px-3 py-1.5 rounded-lg text-sm font-medium transition-all flex items-center gap-2 ${
+                  upcomingFilter === 'next_month' ? 'bg-teal-50 text-teal-700' : 'text-stone-600 hover:bg-stone-50'
+                }`}
+              >
+                <TrendingUp size={14} /> Next Month
+              </button>
+              <button
+                onClick={() => { setUpcomingFilter(upcomingFilter === 'current_month' ? 'none' : 'current_month'); setSelectedMonth('All Months'); }}
+                className={`px-3 py-1.5 rounded-lg text-sm font-medium transition-all flex items-center gap-2 ${
+                  upcomingFilter === 'current_month' ? 'bg-orange-50 text-orange-700' : 'text-stone-600 hover:bg-stone-50'
+                }`}
+              >
+                <Clock size={14} /> This Month
+              </button>
+            </div>
+          </div>
+
+          {/* ACTION BUTTONS */}
+          <div className="flex gap-3 w-full xl:w-auto justify-end">
+            <button
+              className={`px-4 py-2.5 rounded-xl font-medium text-sm flex items-center gap-2 transition-all ${
+                showLiveLocation ? 'bg-teal-600 text-white shadow-lg shadow-teal-200' : 'bg-white text-stone-600 border border-stone-200 hover:bg-stone-50'
+              }`}
+              onClick={() => setShowLiveLocation(x => !x)}
+            >
+              <Share2 size={16} /> Live Location
+            </button>
+            
+            <div className="bg-white p-1 rounded-xl border border-stone-200 flex">
+              <button
+                onClick={() => setShowMap(false)}
+                className={`px-4 py-1.5 rounded-lg text-sm font-medium flex items-center gap-2 transition-all ${!showMap ? 'bg-stone-100 text-stone-900' : 'text-stone-500 hover:text-stone-700'}`}
+              >
+                <Grid size={16} /> List
+              </button>
+              <button
+                onClick={() => setShowMap(true)}
+                className={`px-4 py-1.5 rounded-lg text-sm font-medium flex items-center gap-2 transition-all ${showMap ? 'bg-stone-100 text-stone-900' : 'text-stone-500 hover:text-stone-700'}`}
+              >
+                <MapIcon size={16} /> Map
+              </button>
+            </div>
+          </div>
+        </div>
+
+        {/* LIST OR MAP VIEW */}
+        {!showMap ? (
+          <>
+            <motion.div 
+              layout
+              className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-8"
+            >
+              <AnimatePresence>
+                {showCards.map((card, idx) => (
+                  <motion.div
+                    layout
+                    initial={{ opacity: 0, y: 20 }}
+                    animate={{ opacity: 1, y: 0 }}
+                    exit={{ opacity: 0, scale: 0.9 }}
+                    transition={{ delay: idx * 0.05 }}
+                    key={card.id || card.name}
+                    className="bg-white rounded-2xl shadow-sm border border-stone-100 overflow-hidden hover:shadow-xl transition-all duration-300 group cursor-pointer flex flex-col h-full"
+                    onMouseEnter={() => card.id && setHoveredCardId(card.id)}
+                    onMouseLeave={() => setHoveredCardId(null)}
+                    onClick={() => setSelectedCard(card)}
+                  >
+                    <div className="relative h-56 overflow-hidden">
+                      <img
+                        src={card.img || 'https://placehold.co/500x300?text=Culture+or+Historical'}
+                        alt={card.name}
+                        className="w-full h-full object-cover group-hover:scale-110 transition-transform duration-700"
+                        onError={(e) => {
+                          const img = e.target as HTMLImageElement;
+                          img.src = 'https://via.placeholder.com/500x300?text=Festival+Image';
+                        }}
+                      />
+                      <div className="absolute inset-0 bg-gradient-to-t from-black/70 via-transparent to-transparent opacity-80"></div>
+                      <div className="absolute bottom-4 left-4 right-4">
+                        <span className={`text-[10px] font-bold uppercase tracking-wider px-2 py-1 rounded-md mb-2 inline-block ${
+                          card.cardType === "festival" ? "bg-orange-500 text-white" :
+                          card.cardType === "culture" ? "bg-purple-500 text-white" :
+                          "bg-amber-600 text-white"
+                        }`}>
+                          {card.cardType === "festival" ? ('type' in card ? card.type.split(' ')[0] : "Festival")
+                            : card.cardType === "culture" ? ('aspect' in card ? card.aspect : "Culture")
+                            : "History"}
+                        </span>
+                        <h3 className="text-xl font-bold text-white font-serif leading-tight">{card.name}</h3>
+                      </div>
+                    </div>
+                    <div className="p-5 flex-grow flex flex-col">
+                      {card.cardType === "festival" && (
+                        <>
+                          <div className="flex items-center gap-4 text-xs font-medium text-stone-500 mb-4">
+                            <span className="flex items-center gap-1.5 bg-stone-50 px-2 py-1 rounded-md"><Calendar size={14} className="text-teal-600" /> {card.month}</span>
+                            <span className="flex items-center gap-1.5 bg-stone-50 px-2 py-1 rounded-md"><MapPin size={14} className="text-red-500" /> {card.location}</span>
+                          </div>
+                          <p className="text-stone-600 text-sm leading-relaxed line-clamp-3 mb-4">{card.desc}</p>
+                        </>
+                      )}
+                      {card.cardType === "culture" && (
+                        <p className="text-stone-600 text-sm leading-relaxed line-clamp-4 mb-4">{card.description}</p>
+                      )}
+                      {card.cardType === "historical" && (
+                        <>
+                          <div className="flex items-center gap-2 text-xs font-medium text-stone-500 mb-3">
+                            <span className="flex items-center gap-1.5 bg-stone-50 px-2 py-1 rounded-md"><MapPin size={14} className="text-red-500" /> {card.location}</span>
+                          </div>
+                          <p className="text-stone-600 text-sm leading-relaxed line-clamp-3 mb-4">{card.description}</p>
+                        </>
+                      )}
+                      
+                      <div className="mt-auto pt-4 border-t border-stone-100 flex justify-between items-center">
+                        <span className="text-xs font-bold text-stone-400 uppercase tracking-wider">Read More</span>
+                        <div className="w-8 h-8 rounded-full bg-stone-100 flex items-center justify-center group-hover:bg-stone-900 group-hover:text-white transition-colors">
+                          <ChevronRight size={16} />
+                        </div>
+                      </div>
+                    </div>
+                  </motion.div>
+                ))}
+              </AnimatePresence>
+            </motion.div>
+            
+            {filteredCards.length > 6 && (
+              <div className="text-center mt-12">
+                <button
+                  onClick={() => setShowMore(prev => !prev)}
+                  className="px-8 py-3 bg-white border border-stone-200 text-stone-800 rounded-full shadow-sm font-bold hover:bg-stone-50 hover:shadow-md transition-all transform hover:-translate-y-1"
+                >
+                  {showMore ? "Show Less" : "Discover More"}
+                </button>
+              </div>
+            )}
+          </>
+        ) : (
+          // --- MAP VIEW ---
+          <motion.div 
+            initial={{ opacity: 0, scale: 0.95 }}
+            animate={{ opacity: 1, scale: 1 }}
+            className="w-full h-[600px] rounded-3xl flex flex-col border border-stone-200 overflow-hidden shadow-2xl relative"
+          >
+            <MapContainer
+              center={initialMapCenter}
+              zoom={5}
+              scrollWheelZoom
+              style={{ flexGrow: 1, zIndex: 1 }}
+              ref={mapRef}
+            >
+              <TileLayer url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
+                attribution='&copy; <a href="http://osm.org/copyright">OpenStreetMap</a> contributors'
+              />
+              {filteredCards
+                .filter(card => card.lat && card.lng)
+                .map(card => (
+                  <Marker
+                    key={card.id}
+                    position={[card.lat, card.lng]}
+                    opacity={hoveredCardId === card.id ? 1.0 : 0.8}
+                    eventHandlers={{
+                      mouseover: () => card.id && setHoveredCardId(card.id),
+                      mouseout: () => setHoveredCardId(null),
+                      click: () => setSelectedCard(card)
                     }}
-                  />
-                  <div className="absolute bottom-0 left-0 w-full bg-gradient-to-t from-black/80 to-transparent p-4">
-                    <span className="text-white text-xs bg-orange-600 px-2 py-1 rounded-full mb-2 inline-block font-semibold">
-                        {card.cardType === "festival" ? ('type' in card ? card.type : "")
-                          : card.cardType === "culture" ? ('aspect' in card ? card.aspect : "")
-                          : card.cardType === "historical" ? ('era' in card ? card.era : "") : ""}
+                  >
+                    <Popup>
+                      <div className="font-sans p-1">
+                        <h3 className="font-bold text-stone-900 mb-1">{card.name}</h3>
+                        <p className="text-xs text-stone-500 mb-2">
+                          {(card.cardType === 'festival' || card.cardType === 'historical') && 'location' in card 
+                            ? card.location 
+                            : ""}
+                        </p>
+                        <button 
+                          onClick={() => setSelectedCard(card)}
+                          className="text-xs bg-stone-900 text-white px-2 py-1 rounded hover:bg-stone-700 w-full"
+                        >
+                          View Details
+                        </button>
+                      </div>
+                    </Popup>
+                  </Marker>
+                ))}
+              {/* Only fly to on SELECT, not hover */}
+              {selectedCard && selectedCard.lat && selectedCard.lng && (
+                <FlyToLocation position={[selectedCard.lat, selectedCard.lng]} />
+              )}
+              {showLiveLocation && <LiveLocationMarker />}
+            </MapContainer>
+            
+            <div className="absolute bottom-6 left-6 z-[400]">
+              <button onClick={handleDownloadMap} className="px-4 py-2 bg-white text-stone-800 rounded-lg shadow-lg font-medium text-sm flex items-center gap-2 hover:bg-stone-50">
+                <Download size={16} /> Download Map
+              </button>
+            </div>
+          </motion.div>
+        )}
+      </div>
+
+      {/* HELPLINE FLOATING BUTTON */}
+      <motion.button
+        initial={{ scale: 0 }}
+        animate={{ scale: 1 }}
+        whileHover={{ scale: 1.1 }}
+        whileTap={{ scale: 0.9 }}
+        onClick={() => setShowHelpline(true)}
+        className="fixed bottom-8 right-8 z-40 bg-gradient-to-r from-orange-500 to-pink-600 text-white p-4 rounded-full shadow-2xl shadow-orange-500/40 flex items-center gap-2 group"
+      >
+        <Phone className="w-6 h-6" />
+        <span className="max-w-0 overflow-hidden group-hover:max-w-xs transition-all duration-300 whitespace-nowrap font-bold">
+          Helpline & Info
+        </span>
+      </motion.button>
+
+      {/* HELPLINE MODAL */}
+      <AnimatePresence>
+        {showHelpline && <HelplineModal onClose={() => setShowHelpline(false)} />}
+      </AnimatePresence>
+
+      {/* CARD DETAIL MODAL */}
+      <AnimatePresence>
+        {selectedCard && (
+          <motion.div 
+            initial={{ opacity: 0 }}
+            animate={{ opacity: 1 }}
+            exit={{ opacity: 0 }}
+            className="fixed inset-0 bg-black/60 backdrop-blur-sm flex items-center justify-center z-50 p-4"
+            onClick={() => setSelectedCard(null)}
+          >
+            <motion.div 
+              initial={{ scale: 0.9, opacity: 0, y: 20 }}
+              animate={{ scale: 1, opacity: 1, y: 0 }}
+              exit={{ scale: 0.9, opacity: 0, y: 20 }}
+              className="bg-white rounded-3xl w-full max-w-2xl shadow-2xl overflow-hidden max-h-[90vh] overflow-y-auto"
+              onClick={e => e.stopPropagation()}
+            >
+              <div className="relative h-64 md:h-80">
+                <img
+                  src={selectedCard?.img || 'https://placehold.co/500x300?text=Culture+or+Historical'}
+                  className="w-full h-full object-cover"
+                  alt={selectedCard?.name}
+                />
+                <div className="absolute inset-0 bg-gradient-to-t from-black/80 via-transparent to-transparent"></div>
+                <div className="absolute bottom-0 left-0 p-8">
+                  <h2 className="text-3xl md:text-4xl font-bold text-white font-serif mb-2">{selectedCard?.name}</h2>
+                  <div className="flex flex-wrap gap-2">
+                    {selectedCard?.cardType === "festival" && (
+                      <span className="bg-white/20 backdrop-blur-md text-white px-3 py-1 rounded-full text-sm border border-white/30">
+                        {selectedCard.type}
                       </span>
-                    <h3 className="text-2xl font-bold text-white font-serif">{card.name}</h3>
+                    )}
                   </div>
                 </div>
-                <div className="p-5">
-                  {card.cardType === "festival" && (
+                <button 
+                  onClick={() => setSelectedCard(null)} 
+                  className="absolute top-4 right-4 bg-black/20 backdrop-blur-md text-white p-2 rounded-full hover:bg-black/40 transition"
+                >
+                  <X size={24} />
+                </button>
+              </div>
+              
+              <div className="p-8">
+                <div className="grid grid-cols-1 md:grid-cols-3 gap-6 mb-8">
+                  {selectedCard?.cardType === "festival" && (
                     <>
-                      <div className="flex items-center gap-4 text-sm text-stone-500 mb-3">
-                        <span className="flex items-center gap-1"><Calendar size={14} className="text-teal-600" /> {card.month}</span>
-                        <span className="flex items-center gap-1"><MapPin size={14} className="text-red-500" /> {card.location}</span>
+                      <div className="bg-stone-50 p-4 rounded-xl">
+                        <span className="text-xs font-bold text-stone-400 uppercase block mb-1">When</span>
+                        <span className="font-medium text-stone-800">{selectedCard.month}</span>
                       </div>
-                      <p className="text-stone-600 text-sm mb-4 line-clamp-2">{card.desc}</p>
+                      <div className="bg-stone-50 p-4 rounded-xl col-span-2">
+                        <span className="text-xs font-bold text-stone-400 uppercase block mb-1">Where</span>
+                        <span className="font-medium text-stone-800">{selectedCard.location}</span>
+                      </div>
                     </>
                   )}
-                  {card.cardType === "culture" && (
-                    <p className="text-stone-600 text-sm mb-4">{card.description}</p>
-                  )}
-                  {card.cardType === "historical" && (
+                  {selectedCard?.cardType === "historical" && (
                     <>
-                      <div className="flex items-center gap-2 text-sm text-stone-500 mb-2">
-                        <span><MapPin size={14} className="text-red-500" /> {card.location}</span>
+                      <div className="bg-stone-50 p-4 rounded-xl">
+                        <span className="text-xs font-bold text-stone-400 uppercase block mb-1">Era</span>
+                        <span className="font-medium text-stone-800">{selectedCard.era}</span>
                       </div>
-                      <p className="text-stone-600 text-sm mb-4 line-clamp-2">{card.description}</p>
+                      <div className="bg-stone-50 p-4 rounded-xl col-span-2">
+                        <span className="text-xs font-bold text-stone-400 uppercase block mb-1">Location</span>
+                        <span className="font-medium text-stone-800">{selectedCard.location}</span>
+                      </div>
                     </>
                   )}
                 </div>
-              </div>
-            ))}
-          </div>
-          {filteredCards.length > 6 && (
-            <div className="text-center mt-6">
-              <button
-                onClick={() => setShowMore(prev => !prev)}
-                className="px-6 py-2 bg-orange-600 text-white rounded-lg shadow font-bold"
-              >
-                {showMore ? "Show Less" : "See More"}
-              </button>
-            </div>
-          )}
-        </>
-      ) : (
-        // --- MAP VIEW ---
-        <div className="w-full h-[600px] rounded-2xl flex flex-col border border-stone-300 overflow-hidden shadow-xl">
-          <MapContainer
-            center={initialMapCenter}
-            zoom={5}
-            scrollWheelZoom
-            style={{ flexGrow: 1, zIndex: 1 }}
-            ref={mapRef}
-          >
-            <TileLayer url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
-              attribution='&copy; <a href="http://osm.org/copyright">OpenStreetMap</a> contributors'
-            />
-            {filteredCards
-              .filter(card => card.lat && card.lng)
-              .map(card => (
-                <Marker
-                  key={card.id}
-                  position={[card.lat, card.lng]}
-                  opacity={hoveredCardId === card.id ? 1.0 : 0.8}
-                  eventHandlers={{
-                    mouseover: () => card.id && setHoveredCardId(card.id),
-                    mouseout: () => setHoveredCardId(null),
-                    click: () => setSelectedCard(card)
-                  }}
-                >
-                  <Popup>
-                    <div className="font-sans">
-                      <h3 className="font-bold text-teal-700 mb-1">{card.name}</h3>
-                      <p className="text-xs text-stone-600 italic">
-                        {(card.cardType === 'festival' || card.cardType === 'historical') && 'location' in card 
-                          ? card.location 
-                          : ""}
-                      </p>
-                      <p className="text-sm mt-1">
-                        {card.cardType === 'festival' && 'desc' in card
-                          ? (card.desc as string).substring(0, 50)
-                          : 'description' in card
-                          ? (card.description as string).substring(0, 50)
-                          : ""}...
-                      </p>
-                    </div>
-                  </Popup>
-                </Marker>
-              ))}
-            {/* Only fly to on SELECT, not hover */}
-            {selectedCard && selectedCard.lat && selectedCard.lng && (
-              <FlyToLocation position={[selectedCard.lat, selectedCard.lng]} />
-            )}
-            {showLiveLocation && <LiveLocationMarker />}
-          </MapContainer>
-          <div className="p-4 bg-stone-100 border-t border-stone-200 flex justify-between items-center">
-            <div className="flex items-center gap-3 text-stone-600">
-              <Download size={20} className="text-blue-500" />
-              <button onClick={handleDownloadMap} className="px-4 py-2 bg-blue-500 text-white rounded ml-2">
-                Download Map
-              </button>
-            </div>
-            <button onClick={() => setShowMap(false)} className="text-teal-700 font-medium hover:underline text-sm">Return to List</button>
-          </div>
-        </div>
-      )}
 
-      {/* VIEW TOGGLE BUTTONS */}
-      <div className="flex mt-10 mb-2 justify-center">
-        <div className="bg-stone-100 p-1 rounded-lg flex shadow-inner">
-          <button
-            onClick={() => setShowMap(false)}
-            className={`px-4 py-2 rounded-lg text-sm font-medium flex items-center gap-2 transition-all ${!showMap ? 'bg-white shadow-md text-teal-700' : 'text-stone-500 hover:bg-stone-200'}`}
-          >
-            <Grid size={16} /> List View
-          </button>
-          <button
-            onClick={() => setShowMap(true)}
-            className={`px-4 py-2 rounded-lg text-sm font-medium flex items-center gap-2 transition-all ${showMap ? 'bg-white shadow-md text-teal-700' : 'text-stone-500 hover:bg-stone-200'}`}
-          >
-            <MapIcon size={16} /> Map View
-          </button>
-        </div>
-      </div>
+                <div className="prose prose-stone max-w-none">
+                  <h3 className="text-xl font-bold text-stone-900 mb-4">About</h3>
+                  <p className="text-stone-600 leading-relaxed text-lg">
+                    {selectedCard?.cardType === "festival" ? selectedCard.desc :
+                     selectedCard?.cardType === "culture" ? selectedCard.description :
+                     selectedCard?.cardType === "historical" ? selectedCard.description : ""}
+                  </p>
+                </div>
 
-      {/* CARD MODAL */}
-      {selectedCard && (
-        <div className="fixed inset-0 bg-black/50 flex items-center justify-center z-50 p-4 backdrop-blur-sm">
-          <div className="bg-white rounded-2xl w-full max-w-lg shadow-2xl overflow-hidden animate-in zoom-in-95 duration-200">
-            <div className="relative h-32">
-              <img
-                src={selectedCard?.img || 'https://placehold.co/500x300?text=Culture+or+Historical'}
-                className="w-full h-full object-cover"
-                alt={selectedCard?.name}
-              />
-              <div className="absolute inset-0 bg-black/40 flex items-center p-6">
-                <h2 className="text-2xl font-bold text-white font-serif">{selectedCard?.name}</h2>
+                <div className="mt-8 pt-8 border-t border-stone-100 flex justify-between items-center">
+                  <button className="text-stone-500 hover:text-stone-800 font-medium text-sm flex items-center gap-2">
+                    <Share2 size={16} /> Share
+                  </button>
+                  <button 
+                    onClick={() => setSelectedCard(null)}
+                    className="px-6 py-3 bg-stone-900 text-white rounded-xl font-bold hover:bg-stone-800 transition shadow-lg shadow-stone-200"
+                  >
+                    Close Details
+                  </button>
+                </div>
               </div>
-              <button onClick={() => setSelectedCard(null)} className="absolute top-4 right-4 bg-black/20 text-white p-1 rounded-full hover:bg-black/40 transition">
-                <X size={20} />
-              </button>
-            </div>
-            <div className="p-6">
-              <div className="text-sm">
-                {selectedCard?.cardType === "festival" && (
-                  <>
-                    <b>Month:</b> {selectedCard.month}<br />
-                    <b>Location:</b> {selectedCard.location}<br /><br />
-                    {selectedCard.desc}
-                  </>
-                )}
-                {selectedCard?.cardType === "culture" && (
-                  <>
-                    <b>Aspect:</b> {selectedCard.aspect}<br /><br />
-                    {selectedCard.description}
-                  </>
-                )}
-                {selectedCard?.cardType === "historical" && (
-                  <>
-                    <b>Era:</b> {selectedCard.era}<br />
-                    <b>Location:</b> {selectedCard.location}<br /><br />
-                    {selectedCard.description}
-                  </>
-                )}
-              </div>
-            </div>
-            <div className="bg-stone-50 p-4 border-t border-stone-100 flex justify-end">
-              <button onClick={() => setSelectedCard(null)} className="px-4 py-2 bg-stone-200 text-stone-800 rounded-lg text-sm font-medium hover:bg-stone-300 transition">
-                Close
-              </button>
-            </div>
-          </div>
-        </div>
-      )}
+            </motion.div>
+          </motion.div>
+        )}
+      </AnimatePresence>
     </div>
   );
 };
