@@ -3,48 +3,19 @@
  * Handles communication with backend /api/mood-analyze endpoint
  */
 
-import { getBackendUrl } from '../config/api';
 import type { MoodAnalyzeRequest, MoodAnalyzeResponse } from '../types/moodAnalyzer';
 
-const DEFAULT_BACKEND_URL = 'http://localhost:3001';
+// Force correct backend URL
+const BACKEND_URL = import.meta.env.VITE_BACKEND_URL || 'http://localhost:3001';
 const MOOD_ANALYZE_ENDPOINT = '/api/mood-analyze';
 
-const normalizeBaseUrl = (value: string): string => value.replace(/\/+$/, '');
-
-const resolveBackendOrigin = (): string => {
-  const candidates = [
-    import.meta.env.VITE_BACKEND_URL?.trim(),
-    getBackendUrl(),
-    DEFAULT_BACKEND_URL,
-  ].filter((candidate): candidate is string => Boolean(candidate && candidate.trim()));
-
-  const origin = candidates[0] ?? DEFAULT_BACKEND_URL;
-  return normalizeBaseUrl(origin);
+const buildEndpointUrl = (): string => {
+  // Remove trailing slash from backend URL if present
+  const baseUrl = BACKEND_URL.replace(/\/+$/, '');
+  return `${baseUrl}${MOOD_ANALYZE_ENDPOINT}`;
 };
 
-const buildEndpointUrl = (): string => `${resolveBackendOrigin()}${MOOD_ANALYZE_ENDPOINT}`;
-
-const parseJsonSafely = async (response: Response): Promise<any> => {
-  const text = await response.text();
-
-  if (!text) {
-    return null;
-  }
-
-  try {
-    return JSON.parse(text);
-  } catch (error) {
-    console.error('❌ Failed to parse JSON from mood analyzer response:', error);
-    return null;
-  }
-};
-
-const throwForErrorResponse = (response: Response, body: any): never => {
-  const statusMessage = `HTTP ${response.status}: ${response.statusText}`;
-  const errorMessage = body?.error || body?.message || statusMessage;
-  console.error('❌ Mood analyzer API error:', errorMessage);
-  throw new Error(errorMessage);
-};
+export const getMoodAnalyzerUrl = (): string => buildEndpointUrl();
 
 async function postMoodAnalysis(payload: { imageData?: string; imageUrl?: string }): Promise<MoodAnalyzeResponse> {
   const url = buildEndpointUrl();
@@ -59,7 +30,7 @@ async function postMoodAnalysis(payload: { imageData?: string; imageUrl?: string
   };
 
   try {
-    console.debug('📨 Calling mood analyzer endpoint:', url);
+    console.debug(' Calling mood analyzer endpoint:', url);
 
     const response = await fetch(url, {
       method: 'POST',
@@ -70,36 +41,36 @@ async function postMoodAnalysis(payload: { imageData?: string; imageUrl?: string
       body: JSON.stringify(requestBody),
     });
 
-    const data = await parseJsonSafely(response);
+    const text = await response.text();
+    let data;
+    try {
+      data = JSON.parse(text);
+    } catch (e) {
+      console.error(' Failed to parse JSON:', text);
+      throw new Error(`Invalid JSON response from server: ${text.substring(0, 100)}...`);
+    }
 
     if (!response.ok) {
-      throwForErrorResponse(response, data);
+      const errorMessage = data?.error || data?.message || `HTTP ${response.status}: ${response.statusText}`;
+      console.error(' Mood analyzer API error:', errorMessage);
+      throw new Error(errorMessage);
     }
 
     if (!data) {
-      console.error('❌ Mood analyzer returned empty response body');
       throw new Error('Mood analyzer returned an empty response.');
     }
 
-    console.debug('✅ Mood analyzer response received');
+    console.debug(' Mood analyzer response received');
     return data as MoodAnalyzeResponse;
   } catch (error) {
+    console.error(' Mood analyzer request failed:', error);
     if (error instanceof TypeError && error.message.includes('Failed to fetch')) {
-      console.error('❌ Network error while reaching mood analyzer at', url);
-      throw new Error(`Failed to reach mood analyzer service at ${url}. Is the backend running?`);
+      throw new Error(`Connection refused to ${url}. Is the backend running on port 3001?`);
     }
-
-    console.error('❌ Unexpected mood analyzer client error:', error);
-    throw error instanceof Error ? error : new Error('Unknown error during mood analysis request.');
+    throw error;
   }
 }
 
-export const getMoodAnalyzerUrl = (): string => buildEndpointUrl();
-
 export async function analyzeMoodWithImage(imageData: string): Promise<MoodAnalyzeResponse> {
   return postMoodAnalysis({ imageData });
-}
-
-export async function analyzeMoodWithURL(imageUrl: string): Promise<MoodAnalyzeResponse> {
-  return postMoodAnalysis({ imageUrl });
 }
